% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeAndPlotretval.R
\name{AnalyzeAndPlotretval}
\alias{AnalyzeAndPlotretval}
\title{Evaluate and plot the return value from \code{CombineChainsAndExtractSigs}.}
\usage{
AnalyzeAndPlotretval(
  retval,
  input.catalog,
  out.dir = NULL,
  verbose = TRUE,
  overwrite = TRUE,
  diagnostic.plot = TRUE
)
}
\arguments{
\item{retval}{The output from \code{CombineChainsAndExtractSigs}.}

\item{input.catalog}{input catalog matrix or
path to file with input catalog}

\item{out.dir}{Directory that will be created for the output, including
csv files and plots (pdfs) of extracted signatures and their exposures.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}

\item{overwrite}{If \code{TRUE} overwrite \code{out.dir}
if it exists, otherwise raise an error.}

\item{diagnostic.plot}{If \code{TRUE} plot diagnostic plots
in a subdirectory \code{Diagnostic_Plots} of \code{out.dir}.}
}
\description{
Evaluate and plot the return value from \code{CombineChainsAndExtractSigs}.
}
