% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Burnin.R
\name{Burnin}
\alias{Burnin}
\title{Run the Gibbs sampling burnin (one thread)}
\usage{
Burnin(
  hdp.state,
  seedNumber = 1,
  burnin = 5000,
  cpiter = 3,
  burnin.verbosity = 0,
  burnin.multiplier = 2,
  checkpoint = TRUE
)
}
\arguments{
\item{hdp.state}{An \code{\link[hdpx]{hdpState-class}} object or a list
representation of an \code{\link[hdpx]{hdpState-class}} object.}

\item{seedNumber}{Set the random seed to this value.}

\item{burnin}{The number of burn-in iterations in
one batch. The total number of burnin iterations is
\code{burnin * burnin.multiplier}.}

\item{cpiter}{The number of iterations of concentration
parameter sampling
to perform after each main Gibbs-sample iteration. (See Teh et al.,
"Hierarchical Dirichlet Processes", Journal of the American Statistical
Association 2006;101(476):1566-1581
(https://doi.org/10.1198/016214506000000302).)}

\item{burnin.verbosity}{Verbosity of message statements.}

\item{burnin.multiplier}{Run \code{burnin.multiplier} rounds of \code{burnin}
iterations.
If \code{checkpoint} is \code{TRUE}, save the burnin chain (see
parameter \code{checkpoint}.)
The diagnostic plot \code{diagnostics.likelihood.pdf} can help
determine if the chain is stationary.
The burnin can be continued from a checkpoint file
with \code{\link{ExtendBurnin}} (see argument \code{checkpoint}).}

\item{checkpoint}{If \code{TRUE}, a checkpoint
file called mSigHdp.burnin.checkpoint.\emph{seeddNumber}.Rdata.
will be created
in the current working directory.}
}
\value{
A list with 2 elements: \describe{
\item{\code{hdplist}}{A list representation of
an \code{\link[hdpx]{hdpState-class}} object.}
\item{\code{likelihood}}{A numeric vector with the likelihood at each iteration.}
}
}
\description{
Run the Gibbs sampling burnin (one thread)
}
