% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BurninIteration.R
\name{BurninIteration}
\alias{BurninIteration}
\title{Run hdp extraction and attribution on a spectra catalog file
A function to do burn-in iteration only. This returns a list of hdp object.
This needs to be converted to a hdpState object before hdp_posterior
(hdpx:::as.hdpState(hdplist))}
\usage{
BurninIteration(
  input.catalog,
  seedNumber = 1,
  K.guess,
  multi.types = FALSE,
  verbose = TRUE,
  post.burnin = 4000,
  post.cpiter = 3,
  post.verbosity = 0,
  gamma.alpha = 1,
  gamma.beta = 1
)
}
\arguments{
\item{input.catalog}{Input spectra catalog as a matrix or
in \code{\link[ICAMS]{ICAMS}} format.}

\item{seedNumber}{An integer that is used to generate separate
random seeds for each call to \code{\link[hdpx]{dp_activate}},
and each call of \code{\link[hdpx]{hdp_posterior}}; please see the code
on how this is done. But repeated calls with same value of
\code{seedNumber} and other inputs should produce the same results.}

\item{K.guess}{Suggested initial value of the number of
signatures, passed to \code{\link[hdpx]{dp_activate}} as
\code{initcc}.}

\item{multi.types}{A logical scalar or
a character vector.
If \code{FALSE}, The HDP analysis
will regard all input spectra as one tumor type.

If \code{TRUE}, the HDP analysis
will infer tumor types based on the string before "::" in their names.
e.g. tumor type for "SA.Syn.Ovary-AdenoCA::S.500" would be "SA.Syn.Ovary-AdenoCA"

If \code{multi.types} is a character vector, then it should be of the same length
as the number of columns in \code{input.catalog}, and each value is the
name of the tumor type of the corresponding column in \code{input.catalog}.

e.g. \code{c("SA.Syn.Ovary-AdenoCA", "SA.Syn.Kidney-RCC")}.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}

\item{post.burnin}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{burnin}.}

\item{post.cpiter}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{cpiter}.}

\item{post.verbosity}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{verbosity}.}

\item{gamma.alpha}{Shape parameter of gamma distribution from which
the Dirichlet process concentration parameters are drawn; in this
function the gamma distributions for all Dirichlet processes are the same.}

\item{gamma.beta}{Inverse scale parameter (rate parameter) of gamma distribution
from which the Dirichlet process concentration parameters are drawn; in this
function the gamma distributions for all Dirichlet processes are the same.}
}
\value{
A list with hdp object after burn-in iteration and likelihood of iteration
}
\description{
Run hdp extraction and attribution on a spectra catalog file
A function to do burn-in iteration only. This returns a list of hdp object.
This needs to be converted to a hdpState object before hdp_posterior
(hdpx:::as.hdpState(hdplist))
}
