% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChainBurnin.R
\name{ChainBurnin}
\alias{ChainBurnin}
\title{Prepare an \code{\link[hdpx]{hdpState-class}} object and run the Gibbs sampling burnin.}
\usage{
ChainBurnin(
  hdp.state,
  seedNumber = 1,
  burnin = 5000,
  cpiter = 3,
  burnin.verbosity = 0,
  burnin.multiplier = 2,
  burnin.checkpoint = TRUE
)
}
\arguments{
\item{hdp.state}{An \code{\link[hdpx]{hdpState-class}} object or a list representation of an \code{\link[hdpx]{hdpState-class}} object.}

\item{seedNumber}{An integer that is used to generate separate
random seeds for the call to \code{\link[hdpx]{dp_activate}},
and before the call of \code{\link[hdpx]{hdp_burnin}}.}

\item{burnin}{Pass to \code{\link[hdpx]{hdp_burnin}}
\code{burnin}. The number of burn-in iterations}

\item{cpiter}{Pass to \code{\link[hdpx]{hdp_burnin}}
\code{cpiter}. The number of iterations of concentration parameter sampling
to perform after each iteration.}

\item{burnin.verbosity}{Pass to \code{\link[hdpx]{hdp_burnin}}
\code{verbosity}.Verbosity of debugging statements.
#'}

\item{burnin.multiplier}{A checkpoint setting. \code{burnin.multiplier} rounds of \code{burnin} iterations will be run.
After each round, a burn-in chain will be save for checkpoint.
A total number of 10,000 iterations is recommended for most analysis. However, number of iterations
can be adjusted based on the size of dataset. According to our experience, 20,000 iterations are
needed when analyzing all PCAWG7 genomes (2,780 samples).
The burnin can be continued from a checkpoint file with \code{\link{ExtendBurnin}}.}

\item{burnin.checkpoint}{If \code{TRUE}, a checkpoint for burnin will be created.}
}
\value{
A list with 2 elements: \describe{
\item{\code{hdplist}}{A list representation of
an \code{\link[hdpx]{hdpState-class}} object.}
\item{likelihood}{A numeric vector with the likelihood at each iteration.}
}
}
\description{
Prepare an \code{\link[hdpx]{hdpState-class}} object and run the Gibbs sampling burnin.
}
