% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChainsDiagnosticPlot_test.R
\name{ChainsDiagnosticPlotMo}
\alias{ChainsDiagnosticPlotMo}
\title{Diagnostic plot for a hdpSampleMulti object}
\usage{
ChainsDiagnosticPlotMo(retval, input.catalog, out.dir, verbose)
}
\arguments{
\item{retval}{output from CombinePosteriorChains.A list with the following elements:\describe{
\item{signature}{The extracted signature profiles as a matrix;
rows are mutation types, columns are
samples (e.g. tumors).}
\item{exposure}{The inferred exposures as a matrix of mutation counts;
rows are signatures, columns are samples (e.g. tumors).}

\item{multi.chains}{A \code{\link[hdpx]{hdpSampleMulti-class}} object.
This object has the method \code{\link[hdpx]{chains}} which returns
a list of \code{\link[hdpx]{hdpSampleChain-class}} objects. Each of these
sample chains objects has a method \code{\link[hdpx]{final_hdpState}}
(actually the methods seems to be just \code{hdp})
that returns the \code{hdpState} from which it was generated.}
}}

\item{input.catalog}{ground truth catalog}

\item{out.dir}{Directory that will be created for the output;
if \code{overwrite} is \code{FALSE} then
abort if \code{out.dir} already exits.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}
}
\description{
Diagnostic plot for a hdpSampleMulti object
}
