% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanChlist.R
\name{CleanChlist}
\alias{CleanChlist}
\title{If the job of Gibbs sampling from \code{MultipleSetupAndPosterior}
has an error caught by R, the corresponding element
of chlist has class try-error.
If the job is stopped with, e.g. a segfault,
the \code{chlist} element is NULL.}
\usage{
CleanChlist(chlist, verbose = FALSE)
}
\arguments{
\item{chlist}{A list of \code{\link[hdpx]{hdpSampleChain-class}} objects.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}
}
\value{
Invisibly, the clean, non-error \code{chlist}
This is a list of \code{\link[hdpx]{hdpSampleChain-class}} objects.
}
\description{
If the job of Gibbs sampling from \code{MultipleSetupAndPosterior}
has an error caught by R, the corresponding element
of chlist has class try-error.
If the job is stopped with, e.g. a segfault,
the \code{chlist} element is NULL.
}
