% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineChainsAndExtractSigs.R
\name{CombineChainsAndExtractSigs}
\alias{CombineChainsAndExtractSigs}
\title{Extract signatures, exposures, and other information from multiple posterior sample chains.}
\usage{
CombineChainsAndExtractSigs(
  clean.chlist,
  input.catalog,
  verbose = TRUE,
  high.confidence.prop = 0.9,
  hc.cutoff = 0.1
)
}
\arguments{
\item{clean.chlist}{A list of \code{\link[hdpx]{hdpSampleChain-class}}
objects, typically returned from \code{MultipleSetupAndPosterior}.
Each element is the result of one posterior sample chain.}

\item{input.catalog}{Input spectra catalog as a matrix.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}

\item{high.confidence.prop}{Raw clusters of mutations
found in \eqn{>=} \code{high.confidence.prop} proportion of posterior
samples are signatures with high confidence.
(Passed to \code{\link[hdpx]{interpret_components}}).}

\item{hc.cutoff}{The cutoff of
height of the hierarchical clustering dendrogram used in combining
raw clusters of mutations into aggregated clusters.
(Passed to \code{\link[hdpx]{extract_components_from_clusters}}.)}
}
\value{
Invisibly, a list with the following elements:\describe{
\item{signature}{The extracted signature profiles as a matrix;
rows are mutation types, columns are signatures with
high confidence based on argument
\code{high.confidence.prop}.}

\item{signature.post.samp.number}{A data frame with two columns. The first
column contains the indices
of columns in \code{signature}.
The second columns contains the number
of posterior Gibbs samples that found a raw cluster
contributing to the corresponding signature.}

\item{signature.cdc}{
A numeric data frame.
Each column corresponds to a column in \code{signature}.
Each row represents a biological sample or
MO CHECK an internal Dirichlet process node.
The column sum is the count of all mutations
contributing to corresponding signature.}

\item{exposureProbs}{The inferred exposures as a matrix
of mutation probabilities;
rows are signatures, columns are samples (e.g. tumors). This is
similar to \code{signature.cdc}, but every column was normalized
to sum to 1.}

\item{low.confidence.signature}{The profiles of signatures extracted
with low confidence as a matrix; rows are mutation types,
columns are signatures found in <
\code{high.confidence.prop} of posterior samples.}

\item{low.confidence.post.samp.number}{
Analogous to \code{signature.post.samp.number}, except for
signatures in \code{low.confidence.signature}.}

\item{low.confidence.cdc}{Analogous to
\code{signature.cdc}, except for signatures
in \code{low.confidence.signature}.}

\item{extracted.retval}{A list object returned from
code{\link[hdpx]{extract_components_from_clusters}}.}

}
}
\description{
Extract signatures, exposures, and other information from multiple posterior sample chains.
}
