% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombinePosteriorChains.R
\name{CombinePosteriorChains}
\alias{CombinePosteriorChains}
\title{Extract components and exposures from multiple posterior sample chains}
\usage{
CombinePosteriorChains(
  clean.chlist,
  input.catalog,
  multi.types,
  cluster.method = "kmedians",
  verbose = TRUE,
  cos.merge = 0.9,
  min.sample = 1
)
}
\arguments{
\item{clean.chlist}{It collects the output of multiple independent hdp_posterior calls.}

\item{input.catalog}{Input spectra catalog as a matrix or
in \code{\link[ICAMS]{ICAMS}} format.}

\item{multi.types}{A logical scalar or
a character vector.
If \code{FALSE}, The HDP analysis
will regard all input spectra as one tumor type.

If \code{TRUE}, the HDP analysis
will infer tumor types based on the string before "::" in their names.
e.g. tumor type for "SA.Syn.Ovary-AdenoCA::S.500" would be "SA.Syn.Ovary-AdenoCA"

If \code{multi.types} is a character vector, then it should be of the same length
as the number of columns in \code{input.catalog}, and each value is the
name of the tumor type of the corresponding column in \code{input.catalog}.

e.g. \code{c("SA.Syn.Ovary-AdenoCA", "SA.Syn.Kidney-RCC")}.}

\item{cluster.method}{A \code{kccaFamily} object input for \code{flexclust} package}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}

\item{cos.merge}{The cosine similarity threshold for merging raw clusters
from the posterior sampling chains into "components" i.e. signatures;
passed to \code{\link[hdpx]{hdp_extract_components}}.}

\item{min.sample}{A "component" (i.e. signature) must have at least
this many samples; passed to \code{\link[hdpx]{hdp_extract_components}}.}
}
\value{
Invisibly, a list with the following elements:\describe{
\item{signature}{The extracted signature profiles as a matrix;
rows are mutation types, columns are
samples (e.g. tumors).}
\item{exposure}{The inferred exposures as a matrix of mutation counts;
rows are signatures, columns are samples (e.g. tumors).}

\item{multi.chains}{A \code{\link[hdpx]{hdpSampleMulti-class}} object.
This object has the method \code{\link[hdpx]{chains}} which returns
a list of \code{\link[hdpx]{hdpSampleChain-class}} objects. Each of these
sample chains objects has a method \code{\link[hdpx]{final_hdpState}}
(actually the methods seems to be just \code{hdp})
that returns the \code{hdpState} from which it was generated.}
}
}
\description{
Extract components and exposures from multiple posterior sample chains
}
