% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComponentDiagnosticPlotting.R
\name{ComponentDiagnosticPlotting}
\alias{ComponentDiagnosticPlotting}
\title{Diagnostic plot for a hdpSampleMulti object. This function is
compatible with the return object from Liu's extract_components_from_clusters}
\usage{
ComponentDiagnosticPlotting(retval, input.catalog, out.dir, verbose)
}
\arguments{
\item{retval}{Return from \code{\link{CombineChainsAndExtractSigs}}}

\item{input.catalog}{Input spectra catalog as a matrix or
in \code{\link[ICAMS]{ICAMS}} format.}

\item{out.dir}{Directory that will be created for the output;
if \code{overwrite} is \code{FALSE} then
abort if \code{out.dir} already exits.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}
}
\description{
Diagnostic plot for a hdpSampleMulti object. This function is
compatible with the return object from Liu's extract_components_from_clusters
}
