% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtendBurnin.R
\name{ExtendBurnin}
\alias{ExtendBurnin}
\title{Extend burnin iterations generated from \code{\link{Burnin}}}
\usage{
ExtendBurnin(hdplist, seedNumber = 1, burnin = 4000, cpiter = 3, verbosity = 0)
}
\arguments{
\item{hdplist}{A list representation of
an \code{\link[hdpx]{hdpState-class}} object;
a return value \code{\link{Burnin}}.}

\item{seedNumber}{A random seed for reproducible results.}

\item{burnin}{The number of burnin iterations to perform.}

\item{cpiter}{The number of iterations of concentration
parameter sampling
to perform after each main Gibbs-sample iteration. (See Teh et al.,
"Hierarchical Dirichlet Processes", Journal of the American Statistical
Association 2006;101(476):1566-1581
(https://doi.org/10.1198/016214506000000302).)}

\item{verbosity}{Passed to \code{\link[hdpx]{hdp_posterior}}
\code{verbosity}.}
}
\value{
The same type of object as returned from
\code{\link{Burnin}.}

The envisioned application is extending burnins from burnin checkpoints.
}
\description{
Extend burnin iterations generated from \code{\link{Burnin}}
}
