% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateAverageCluster.R
\name{GenerateAverageCluster}
\alias{GenerateAverageCluster}
\title{Generate average pattern of clusters of each posterior chain
from combined list of multiple posterior sample chains}
\usage{
GenerateAverageCluster(clean.chlist)
}
\arguments{
\item{clean.chlist}{A list of multiple (or one) posterior sample chains.}
}
\value{
A list of matrices containing the average pattern of clusters within each posterior chain
and a list of matrices containing the sum of each cluster in each posterior chain
}
\description{
Generate average pattern of clusters of each posterior chain
from combined list of multiple posterior sample chains
}
