% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewRunHdpParallel.R
\name{NewRunHdpParallel}
\alias{NewRunHdpParallel}
\title{Extract mutational signatures and optionally compare them to existing signatures and exposures.}
\usage{
NewRunHdpParallel(
  input.catalog,
  seedNumber = 1,
  K.guess,
  multi.types = FALSE,
  verbose = TRUE,
  post.burnin = 4000,
  post.n = 50,
  post.space = 50,
  post.cpiter = 3,
  post.verbosity = 0,
  CPU.cores = 1,
  num.child.process = 4,
  cos.merge = 0.9,
  confident.prop = 0.9,
  noise.prop = 0.5,
  hc.cutoff = 0.1,
  ground.truth.sig = NULL,
  ground.truth.exp = NULL,
  overwrite = TRUE,
  out.dir = NULL,
  gamma.alpha = 1,
  gamma.beta = 1,
  gamma0.alpha = gamma.alpha,
  gamma0.beta = gamma.beta,
  checkpoint.chlist = TRUE,
  checkpoint.1.chain = TRUE,
  prior.sigs = NULL,
  prior.pseudoc = NULL,
  burnin.multiplier = 1,
  burnin.checkpoint = FALSE
)
}
\arguments{
\item{input.catalog}{Input spectra catalog as a matrix or
in \code{\link[ICAMS]{ICAMS}} format.}

\item{seedNumber}{A random seeds passed to \code{\link[hdpx]{dp_activate}}.}

\item{K.guess}{Suggested initial value of the number of
signatures, passed to \code{\link[hdpx]{dp_activate}} as
\code{initcc}.}

\item{multi.types}{A logical scalar or
a character vector.
If \code{FALSE}, The HDP analysis
will regard all input spectra as one tumor type.

If \code{TRUE}, the HDP analysis
will infer tumor types based on the string before "::" in their names.
e.g. tumor type for "SA.Syn.Ovary-AdenoCA::S.500" would be "SA.Syn.Ovary-AdenoCA"

If \code{multi.types} is a character vector, then it should be of the same length
as the number of columns in \code{input.catalog}, and each value is the
name of the tumor type of the corresponding column in \code{input.catalog}.

e.g. \code{c("SA.Syn.Ovary-AdenoCA", "SA.Syn.Kidney-RCC")}.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}

\item{post.burnin}{Pass to \code{\link[hdpx]{hdp_posterior_sample}}
\code{burnin}.}

\item{post.n}{Pass to \code{\link[hdpx]{hdp_posterior_sample}}
\code{n}.}

\item{post.space}{Pass to \code{\link[hdpx]{hdp_posterior_sample}}
\code{space}.}

\item{post.cpiter}{Pass to \code{\link[hdpx]{hdp_posterior_sample}}
\code{cpiter}.}

\item{post.verbosity}{Pass to \code{\link[hdpx]{hdp_posterior_sample}}
\code{verbosity}.}

\item{CPU.cores}{Number of CPUs to use; there is no
point in making this larger than \code{num.child.process}.}

\item{num.child.process}{Number of posterior sampling chains; can set to
1 for testing.}

\item{cos.merge}{The cosine similarity threshold for merging raw clusters
from the posterior sampling chains into "components" i.e. signatures;
passed to \code{\link[hdpx]{extract_sigs_from_clusters}}.}

\item{confident.prop}{Passed to \code{\link[hdpx]{interpret_components}}.
clusters with at least \code{confident.prop} of posterior
samples are high confident signatures}

\item{noise.prop}{Passed to \code{\link[hdpx]{interpret_components}}.
Clusters with less than \code{noise.prop} of posterior samples
are noise signatures}

\item{hc.cutoff}{passed to \code{\link[hdpx]{extract_components_from_clusters}}. The cutoff of
height of hierarchical clustering dendrogram}

\item{ground.truth.sig}{Optional. Either a string with the
path to file with ground truth signatures or and
\code{\link[ICAMS]{ICAMS}} catalog with the ground truth signatures.
These are the signatures used to construct the ground truth spectra.}

\item{ground.truth.exp}{Optional. Ground truth exposure matrix or
path to file with ground truth exposures.
If \code{NULL} skip checks that need this information.}

\item{overwrite}{If \code{TRUE} overwrite \code{out.dir} if it exists, otherwise
raise an error.}

\item{out.dir}{Directory that will be created for the output;
if \code{overwrite} is \code{FALSE} then
abort if \code{out.dir} already exits.}

\item{gamma.alpha}{Shape parameter of
the gamma distribution prior for the Dirichlet process concentration
parameters; in this
function the gamma distributions for all Dirichlet processes,
except possibly the top level process, are the same.}

\item{gamma.beta}{Inverse scale parameter (rate parameter) of
the gamma distribution prior for the Dirichlet process concentration
parameters; in this
function the gamma distributions for all Dirichlet processes, except
possibly the top level process, are the same.}

\item{gamma0.alpha}{See figure B.1 from Nicola Robert's thesis.
The shape parameter (\eqn{\alpha_0}) of the gamma
distribution priors for
the Dirichlet process concentration parameters  (\eqn{\gamma_0})
for \eqn{G_0}.}

\item{gamma0.beta}{See figure B.1 from Nicola Robert's thesis.
Inverse scale parameter (rate parameter, \eqn{\beta_0}) of the gamma
distribution priors for
the Dirichlet process concentration parameters (\eqn{\gamma_0})
for \eqn{G_0}.}

\item{checkpoint.chlist}{If \code{TRUE}, checkpoint the (unclean)
chlist to "initial.chlist.Rdata" in the current working directory.
and checkpoint the clean chlist to "clean.chlist.Rdata" in the
current working directory.}

\item{checkpoint.1.chain}{If \code{TRUE} checkpoint the sample
chain to current working directory, in a file called
sample.chain.\emph{seed_number}.Rdata.}

\item{prior.sigs}{A matrix containing prior signatures.}

\item{prior.pseudoc}{A numeric list. Pesudo counts of each prior signature. Recommended is 1000. In practice,
it may be advisable to put lower weights on prior signatures that you do not expect
to be present in your dataset, or even exclude some priors entirely.}

\item{burnin.multiplier}{A checkpoint setting. \code{burnin.multiplier} rounds of \code{burnin} iterations will be run.
After each round, a burn-in chain will be save for checkpoint.}

\item{burnin.checkpoint}{Default is False. If True, a checkpoint for burnin will be created.}
}
\value{
Invisibly, a list with the following elements:\describe{
\item{signature}{The extracted signature profiles as a matrix;
rows are mutation types, columns are signatures including
high confident signatures -'hdp' signatures and moderate
confident signatures - 'potential hdp' signatures.}

\item{signature.post.samp.number}{A dataframe with two columns. The first
column corresponds to each signature in \code{signature}
and the second columns contains the number of posterior
samples that found the raw clusters contributing to the signature.}

\item{signature.cdc}{A \code{\link[hdpx]{categ_dp_counts}} like dataframe.
Each column corresponds to the sum of all \code{\link[hdpx]{categ_dp_counts}}
matrices of the raw clusters contributing to each signature in code{signature}}

\item{exposureProbs}{The inferred exposures as a matrix of mutation probabilities;
rows are signatures, columns are samples (e.g. tumors).}

\item{noise.signature}{The extracted signature profiles as a matrix; rows are mutation types,
columns are signatures with less than \code{noise.prop} of posterior samples }

\item{noise.post.samp.number}{A data frame with two columns. The first column corresponds
to each signature in \code{noise.signature} and the second
column contains the number of posterior samples that found
the raw clusters contributing to the signature.}

\item{noise.cdc}{A \code{\link[hdpx]{categ_dp_counts}} like data frame. Each column corresponds
to the sum of all \code{\link[hdpx]{categ_dp_counts}} matrices of the raw clusters
contributing to each signature in code{noise.signature}}

\item{extracted.retval}{A list object returned from code{\link[hdpx]{interpret_components}}.}

}
}
\description{
Extract mutational signatures and optionally compare them to existing signatures and exposures.
}
