% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelPosteriorafterBurnin.R
\name{ParallelPosteriorafterBurnin}
\alias{ParallelPosteriorafterBurnin}
\title{Generate an HDP Gibbs sampling chain from a spectra catalog.}
\usage{
ParallelPosteriorafterBurnin(
  retval,
  seedNumber = 1,
  verbose = TRUE,
  post.burnin = 4000,
  post.n = 50,
  post.space = 50,
  post.cpiter = 3,
  post.verbosity = 0,
  num.child.process = 2,
  CPU.cores = 2
)
}
\arguments{
\item{retval}{A list object containing hdplist after burn-in iteration
and likelihood from \code{BurninIteration}.}

\item{seedNumber}{Pass to \code{\link[hdpx]{hdp_posterior}}}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}

\item{post.burnin}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{burnin}. This can be set to a small number}

\item{post.n}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{n}.}

\item{post.space}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{space}.}

\item{post.cpiter}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{cpiter}.}

\item{post.verbosity}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{verbosity}.}

\item{num.child.process}{Number of posterior sampling chains; can set to
1 for testing.}

\item{CPU.cores}{Number of CPUs to use; there is no
point in making this larger than \code{num.child.process}.}
}
\value{
Invisibly, an \code{\link[hdpx]{hdpSampleChain-class}} object
as returned from \code{\link[hdpx]{hdp_posterior}}.
}
\description{
Generate an HDP Gibbs sampling chain from a spectra catalog.
}
