% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepInit.R
\name{PrepInit}
\alias{PrepInit}
\title{Initialize hdp object
Allocate process index for hdp initialization.
Prepare for \code{\link[hdpx]{hdp_init}}}
\usage{
PrepInit(
  multi.types,
  input.catalog,
  verbose = TRUE,
  K.guess,
  gamma.alpha = 1,
  gamma.beta = 1,
  gamma0.alpha = gamma.alpha,
  gamma0.beta = gamma.beta
)
}
\arguments{
\item{multi.types}{A logical scalar or
a character vector.
If \code{FALSE}, The HDP analysis
will regard all input spectra as one tumor type.
HDP structure as one parent node for all tumors

If \code{TRUE}, the HDP analysis
will infer tumor types based on the string before "::" in their names.
e.g. tumor type for "Ovary-AdenoCA::S.500" would be "Ovary-AdenoCA"
HDP structure as a grandparent node for whole data and
one parent node for each tumor type

If \code{multi.types} is a character vector, then it should be of the same length
as the number of columns in \code{input.catalog}, and each value is the
name of the tumor type of the corresponding column in \code{input.catalog}.

If not \code{FALSE}, HDP will give a parent node for each tumor type and a grandparent node
for the whole dataset.

e.g. \code{c("Ovary-AdenoCA", "Kidney-RCC")}.}

\item{input.catalog}{Input spectra catalog as a matrix or
in \code{\link[ICAMS]{ICAMS}} format.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}

\item{K.guess}{Suggested initial value of the number of
clusters. Usually, the number of clusters is two times of the number
of extracted signatures. Passed to \code{\link[hdpx]{dp_activate}} as
\code{initcc}.}

\item{gamma.alpha}{Shape parameter of
the gamma distribution prior for the Dirichlet process concentration
parameters; in this
function the gamma distributions for all Dirichlet processes,
except possibly the top level process, are the same.}

\item{gamma.beta}{Inverse scale parameter (rate parameter) of
the gamma distribution prior for the Dirichlet process concentration
parameters; in this
function the gamma distributions for all Dirichlet processes, except
possibly the top level process, are the same.

We recommend gamma.alpha = 1 and gamma.beta = 20 for single-base-substitution signatures extraction;
gamma.alpha = 1 and gamma.beta = 50  for doublet-base-substitution/INDEL signature extraction}

\item{gamma0.alpha}{See figure B.1 from Nicola Robert's thesis.
The shape parameter (\eqn{\alpha_0}) of the gamma
distribution priors for
the Dirichlet process concentration parameters  (\eqn{\gamma_0})
for \eqn{G_0}.}

\item{gamma0.beta}{See figure B.1 from Nicola Robert's thesis.
Inverse scale parameter (rate parameter, \eqn{\beta_0}) of the gamma
distribution priors for
the Dirichlet process concentration parameters (\eqn{\gamma_0})
for \eqn{G_0}.}
}
\description{
Initialize hdp object
Allocate process index for hdp initialization.
Prepare for \code{\link[hdpx]{hdp_init}}
}
