% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriorSetupAndActivate.R
\name{PriorSetupAndActivate}
\alias{PriorSetupAndActivate}
\title{Generate an HDP Gibbs sampling chain from a spectra catalog.}
\usage{
PriorSetupAndActivate(
  prior.sigs,
  prior.pseudoc,
  gamma.alpha = 1,
  gamma.beta = 20,
  K.guess,
  multi.types = F,
  input.catalog,
  verbose = TRUE,
  seedNumber = 1
)
}
\arguments{
\item{prior.sigs}{DELETE ME LATER, NOT SUPPORTED. A matrix containing prior signatures.}

\item{prior.pseudoc}{DELETE ME LATER, NOT SUPPORTED. A numeric list. Pseudo counts of each prior signature. Recommended is 1000. In practice,
it may be advisable to put lower weights on prior signatures that you do not expect
to be present in your dataset, or even exclude some priors entirely.}

\item{gamma.alpha}{Shape parameter of
the gamma distribution prior for the Dirichlet process concentration
parameters; in this
function the gamma distributions for all Dirichlet processes,
except possibly the top level process, are the same.}

\item{gamma.beta}{Inverse scale parameter (rate parameter) of
the gamma distribution prior for the Dirichlet process concentration
parameters; in this
function the gamma distributions for all Dirichlet processes, except
possibly the top level process, are the same.}

\item{K.guess}{Suggested initial value of the number of
signatures, passed to \code{\link[hdpx]{dp_activate}} as
\code{initcc}.}

\item{multi.types}{A logical scalar or
a character vector.
If \code{FALSE}, The HDP analysis
will regard all input spectra as one tumor type.

If \code{TRUE}, the HDP analysis
will infer tumor types based on the string before "::" in their names.
e.g. tumor type for "SA.Syn.Ovary-AdenoCA::S.500" would be "SA.Syn.Ovary-AdenoCA"

If \code{multi.types} is a character vector, then it should be of the same length
as the number of columns in \code{input.catalog}, and each value is the
name of the tumor type of the corresponding column in \code{input.catalog}.

e.g. \code{c("SA.Syn.Ovary-AdenoCA", "SA.Syn.Kidney-RCC")}.}

\item{input.catalog}{Input spectra catalog as a matrix or
in \code{\link[ICAMS]{ICAMS}} format.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}

\item{seedNumber}{A random seeds passed to \code{\link[hdpx]{dp_activate}}.}
}
\value{
Invisibly, an \code{\link[hdpx]{hdpState-class}} object
as returned from \code{\link[hdpx]{dp_activate}}.
}
\description{
Generate an HDP Gibbs sampling chain from a spectra catalog.
}
\keyword{internal}
