% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runhdp4.R
\name{Runhdp4}
\alias{Runhdp4}
\title{Run hdp extraction and attribution on a spectra catalog file using hdpx}
\usage{
Runhdp4(
  input.catalog,
  out.dir,
  CPU.cores = 1,
  seedNumber = 1,
  K.guess,
  multi.types = FALSE,
  remove.noise = FALSE,
  test.only = 0,
  overwrite = FALSE,
  verbose = TRUE,
  num.posterior = 4,
  post.burnin = 4000,
  post.n = 50,
  post.space = 50,
  post.cpiter = 3,
  post.verbosity = 0,
  cos.merge = 0.9,
  min.sample = 1,
  plot.extracted.sig = FALSE
)
}
\arguments{
\item{input.catalog}{Either a character string, in which case this
is the path to a file containing a spectra catalog
in \code{\link[ICAMS]{ICAMS}} format, or an \code{\link[ICAMS]{ICAMS}} catalog.}

\item{out.dir}{Directory that will be created for the output;
if \code{overwrite} is \code{FALSE} then
abort if \code{out.dir} already exits.}

\item{CPU.cores}{Number of CPUs to use in running
\code{\link[hdp]{hdp_posterior}}; this is used to parallize
running the posterior sampling chains, so there is no
point in making this larger than \code{num.posterior}.}

\item{seedNumber}{An integer that is used to generate separate
random seeds for each call to \code{\link[hdp]{dp_activate}},
and each call of \code{\link[hdp]{hdp_posterior}}; please see the code
on how this is done. But repeated calls with same value of
\code{seedNumber} and other inputs should produce the same results.}

\item{K.guess}{Suggested initial value of the number of
signatures, passed to \code{\link[hdp]{dp_activate}} as
\code{initcc}.}

\item{multi.types}{A logical scalar or
a character vector.
If \code{FALSE}, hdp will regard all input spectra as one tumor type.

If \code{TRUE}, hdp will infer tumor types based on the string before "::" in their names.
e.g. tumor type for "SA.Syn.Ovary-AdenoCA::S.500" would be "SA.Syn.Ovary-AdenoCA"

If \code{multi.types} is a character vector, then it should be of the same length
as the number of columns in \code{input.catalog}, and each value is the
name of the tumor type of the corresponding column in \code{input.catalog},
e.g. \code{c("SA.Syn.Ovary-AdenoCA", "SA.Syn.Ovary-AdenoCA", "SA.Syn.Kidney-RCC")}.}

\item{remove.noise}{Deprecated; ignored}

\item{test.only}{If > 0, only analyze the first \code{test.only} columns
in \code{input.catalog}.}

\item{overwrite}{If \code{TRUE} overwrite \code{out.dir} if it exists, otherwise
raise an error.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}

\item{num.posterior}{Number of posterior sampling chains; can set to
1 for testing.}

\item{post.burnin}{Pass to \code{\link[hdp]{hdp_posterior}}
\code{burnin}.}

\item{post.n}{Pass to \code{\link[hdp]{hdp_posterior}}
\code{n}.}

\item{post.space}{Pass to \code{\link[hdp]{hdp_posterior}}
\code{space}.}

\item{post.cpiter}{Pass to \code{\link[hdp]{hdp_posterior}}
\code{cpiter}.}

\item{post.verbosity}{Pass to \code{\link[hdp]{hdp_posterior}}
\code{verbosity}.}

\item{cos.merge}{The cosine similarity threshhold for merging raw clusters
from the posterior sampling chains into "components" i.e. signatures;
passed to \code{\link[hdp]{hdp_extract_components}}.}

\item{min.sample}{A "component" (i.e. signature) must have at least
this many samples; passed to \code{\link[hdp]{hdp_extract_components}}.}

\item{plot.extracted.sig}{If \code{TRUE} then plot the extracted signatures.}
}
\value{
The same list as returned by \code{\link{RunhdpInternal4}}.
}
\description{
Run hdp extraction and attribution on a spectra catalog file using hdpx
}
\details{
Creates several files in \code{out.dir}. These are:
call.and.session.info.txt, hdp.diagnostics.pdf, Runhdp4.retval.Rdata,
extracted.signatures.csv, extracted.signature.pdf (optional),
inferred.exposures.csv.
}
