% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunHdpParallel.R
\name{RunHdpParallel}
\alias{RunHdpParallel}
\title{A caller function combines MultipleSetupandPosterior, CombinePosteriorChains and AnalyzeAndPlotretval}
\usage{
RunHdpParallel(
  input.catalog,
  seedNumber = 1,
  K.guess,
  multi.types = FALSE,
  verbose = TRUE,
  post.burnin = 4000,
  post.n = 50,
  post.space = 50,
  post.cpiter = 3,
  post.verbosity = 0,
  CPU.cores = 1,
  num.child.process = 4,
  ground.truth.sig,
  ground.truth.exp,
  out.dir = NULL
)
}
\arguments{
\item{input.catalog}{Input spectra catalog as a matrix or
in \code{\link[ICAMS]{ICAMS}} format.}

\item{seedNumber}{An integer that is used to generate separate
random seeds for each call to \code{\link[hdpx]{dp_activate}},
and each call of \code{\link[hdpx]{hdp_posterior}}; please see the code
on how this is done. But repeated calls with same value of
\code{seedNumber} and other inputs should produce the same results.}

\item{K.guess}{Initial guess of the number of "raw clusters",
which may be larger than the number of signatures
(sometimes called "components" in the hdpx code);
passed to \code{\link[hdpx]{dp_activate}} as
\code{initcc}.}

\item{multi.types}{A logical scalar or
a character vector.
If \code{FALSE}, The HDP analysis
will regard all input spectra as one tumor type.

If \code{TRUE}, the HDP analysis
will infer tumor types based on the string before "::" in their names.
e.g. tumor type for "SA.Syn.Ovary-AdenoCA::S.500" would be "SA.Syn.Ovary-AdenoCA"

If \code{multi.types} is a character vector, then it should be of the same length
as the number of columns in \code{input.catalog}, and each value is the
name of the tumor type of the corresponding column in \code{input.catalog},
e.g. \code{c("SA.Syn.Ovary-AdenoCA", "SA.Syn.Ovary-AdenoCA", "SA.Syn.Kidney-RCC")}.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}

\item{post.burnin}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{burnin}.}

\item{post.n}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{n}.}

\item{post.space}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{space}.}

\item{post.cpiter}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{cpiter}.}

\item{post.verbosity}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{verbosity}.}

\item{CPU.cores}{Number of CPUs to use; there is no
point in making this larger than \code{num.child.process}.}

\item{num.child.process}{Number of posterior sampling chains; can set to
1 for testing.}

\item{ground.truth.sig}{Use this one. Either a string with the
path to file with ground truth signatures or and
\code{\link[ICAMS]{ICAMS}} catalog with the ground truth signatures.
These are the signatures used to construct the ground truth spectra.}

\item{ground.truth.exp}{Ground truth exposure matrix or
path to file with ground truth exposures.
If \code{NULL} skip checks that need this information.}

\item{out.dir}{Directory that will be created for the output}
}
\value{
If \code{out.dir} is not NULL, output including data and plots
will be saved in \code{out.dir}.
Else,invisibly, a list with the following elements:\describe{
\item{signature}{The extracted signature profiles as a matrix;
rows are mutation types, columns are
samples (e.g. tumors).}
\item{exposure}{The inferred exposures as a matrix of mutation counts;
rows are signatures, columns are samples (e.g. tumors).}

\item{multi.chains}{A \code{\link[hdpx]{hdpSampleMulti-class}} object.
This object has the method \code{\link[hdpx]{chains}} which returns
a list of \code{\link[hdpx]{hdpSampleChain-class}} objects. Each of these
sample chains objects has a method \code{\link[hdpx]{final_hdpState}}
(actually the methods seems to be just \code{hdp})
that returns the \code{hdpState} from which it was generated.}
}
}
\description{
A caller function combines MultipleSetupandPosterior, CombinePosteriorChains and AnalyzeAndPlotretval
}
