% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetupAndPosterior.R
\name{SetupAndPosterior}
\alias{SetupAndPosterior}
\title{Generate an HDP Gibbs sampling chain from a spectra catalog.}
\usage{
SetupAndPosterior(
  input.catalog,
  seedNumber = 1,
  K.guess,
  multi.types = FALSE,
  verbose = TRUE,
  post.burnin = 4000,
  post.n = 50,
  post.space = 50,
  post.cpiter = 3,
  post.verbosity = 0,
  gamma.alpha = 1,
  gamma.beta = 1,
  one.parent.hack = FALSE
)
}
\arguments{
\item{input.catalog}{Input spectra catalog as a matrix or
in \code{\link[ICAMS]{ICAMS}} format.}

\item{seedNumber}{A random seeds passed to \code{\link[hdpx]{dp_activate}}.}

\item{K.guess}{Suggested initial value of the number of
signatures, passed to \code{\link[hdpx]{dp_activate}} as
\code{initcc}.}

\item{multi.types}{A logical scalar or
a character vector.
If \code{FALSE}, The HDP analysis
will regard all input spectra as one tumor type.

If \code{TRUE}, the HDP analysis
will infer tumor types based on the string before "::" in their names.
e.g. tumor type for "SA.Syn.Ovary-AdenoCA::S.500" would be "SA.Syn.Ovary-AdenoCA"

If \code{multi.types} is a character vector, then it should be of the same length
as the number of columns in \code{input.catalog}, and each value is the
name of the tumor type of the corresponding column in \code{input.catalog}.

e.g. \code{c("SA.Syn.Ovary-AdenoCA", "SA.Syn.Kidney-RCC")}.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}

\item{post.burnin}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{burnin}.}

\item{post.n}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{n}.}

\item{post.space}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{space}.}

\item{post.cpiter}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{cpiter}.}

\item{post.verbosity}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{verbosity}.}

\item{gamma.alpha}{Shape parameter of gamma distribution from which
the Dirichlet process concentration parameters are drawn; in this
function the gamma distributions for all Dirichlet processes are the same.}

\item{gamma.beta}{Inverse scale parameter (rate parameter) of gamma distribution
from which the Dirichlet process concentration parameters are drawn; in this
function the gamma distributions for all Dirichlet processes are the same.}
}
\value{
Invisibly, an \code{\link[hdpx]{hdpSampleChain-class}} object
as returned from \code{\link[hdpx]{hdp_posterior}}.
}
\description{
Generate an HDP Gibbs sampling chain from a spectra catalog.
}
