% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_downsample_curves.R
\name{show_downsample_curves}
\alias{show_downsample_curves}
\title{Show the effects of the downsample_threshold argument to \code{\link{RunHdpxParallel}}.}
\usage{
show_downsample_curves(...)
}
\arguments{
\item{...}{One or more positive numbers to use as
possible values for \code{downsample_threshold}.}
}
\value{
Called for side effects.
}
\description{
Show the effects of the downsample_threshold argument to \code{\link{RunHdpxParallel}}.
}
\examples{
show_downsample_curves(3000, 5000, 10000)

}
