% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotExposure.R
\name{SortExp}
\alias{SortExp}
\title{Sort columns of an exposure matrix from largest to smaller (or vice versa).}
\usage{
SortExp(exposures, decreasing = TRUE)
}
\arguments{
\item{exposures}{The exposures to sort; columns are samples.}

\item{decreasing}{If \code{TRUE} sort from largest to smallest.}
}
\description{
Sort columns of an exposure matrix from largest to smaller (or vice versa).
}
