% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test.ground.truth.sig}
\alias{test.ground.truth.sig}
\alias{test.ground.truth.exposure}
\title{test.ground.truth.sig}
\format{
An \code{\link[ICAMS]{ICAMS}} catalog (each column is a COSMICv3 signature, each row is a mutation type,
e.g. ACT > AGT).

An exposure matrix (each row is a COSMICv3 signature used to generate \code{test.spectra} , each column is a sample).
}
\usage{
test.ground.truth.sig

test.ground.truth.exposure
}
\description{
Ground truth signature for testing.

Ground truth exposure for testing.
}
\keyword{datasets}
