% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test.spectra}
\alias{test.spectra}
\title{test.spectra}
\format{
An \code{\link[ICAMS]{ICAMS}} catalog (each column is a sample, each row is a mutation type,
e.g. ACT > AGT).
}
\usage{
test.spectra
}
\description{
Synthetic SBS696 spectra for testing.
}
\keyword{datasets}
