% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmake.s.R
\name{xmake.s}
\alias{xmake.s}
\title{Return a function to calculate the unsigned Stirling numbers of the first kind}
\usage{
xmake.s()
}
\value{
A function to calculate a vector of unsigned Stirling numbers,
\eqn{s(n ,k), k = 1...n},
each divided by the maximum Stirling number in the series.
The returned function is a closure with state that includes
a list of all the unsigned Stirling number series \eqn{<=} the argument, \eqn{n},

i.e. \eqn{[s(1, 1)], [s(2, 1), s(2, 2)], ..., [s(n, 1), ..., s(n, n)]}.
Memory usage could be substantial, but the stored state
does not include the many trailing zeros in the vectors.
For this to work within the hdp (https://github.com/nicolaroberts/hdp)
package the
function returned \emph{must} be called \code{stir.closure}.
}
\description{
Return a function to calculate the unsigned Stirling numbers of the first kind
}
