% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeAndPlotretval.R
\name{AnalyzeAndPlotretval}
\alias{AnalyzeAndPlotretval}
\title{Evaluate and plot retval from CombinePosteriorChains}
\usage{
AnalyzeAndPlotretval(
  retval,
  out.dir = NUL,
  ground.truth.sig,
  ground.truth.exp = NULL,
  verbose = TRUE,
  overwrite = TRUE
)
}
\arguments{
\item{retval}{the output from function CombinePosteriorChains}

\item{out.dir}{Directory that will be created for the output;
if \code{overwrite} is \code{FALSE} then
abort if \code{out.dir} already exits.}

\item{ground.truth.sig}{Use this one. Either a string with the
path to file with ground truth signatures or and
\code{\link[ICAMS]{ICAMS}} catalog with the ground truth signatures.
These are the signatures used to construct the ground truth spectra.}

\item{ground.truth.exp}{Ground truth exposure matrix or
path to file with ground truth exposures.
If \code{NULL} skip checks that need this information.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}

\item{overwrite}{If \code{TRUE} overwrite \code{out.dir} if it exists, otherwise
raise an error.}
}
\description{
Evaluate and plot retval from CombinePosteriorChains
}
