% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChainBurnin.R
\name{ChainBurnin}
\alias{ChainBurnin}
\title{Prepare an \code{\link[hdpx]{hdpState-class}} object and run the Gibbs sampling burnin.}
\usage{
ChainBurnin(
  hdp.state,
  seedNumber = 1,
  burnin = 4000,
  cpiter = 3,
  burnin.verbosity = 0,
  burnin.multiplier = 1,
  burnin.checkpoint = FALSE
)
}
\arguments{
\item{hdp.state}{An \code{\link[hdpx]{hdpState-class}} object or a list representation of an \code{\link[hdpx]{hdpState-class}} object.}

\item{seedNumber}{An integer that is used to generate separate
random seeds for the call to \code{\link[hdpx]{dp_activate}},
and before the call of \code{\link[hdpx]{hdp_burnin}}.}

\item{burnin}{Pass to \code{\link[hdpx]{hdp_burnin}}
\code{burnin}.}

\item{cpiter}{Pass to \code{\link[hdpx]{hdp_burnin}}
\code{cpiter}.}

\item{burnin.verbosity}{Pass to \code{\link[hdpx]{hdp_burnin}}
\code{verbosity}.}

\item{burnin.multiplier}{A checkpoint setting. \code{burnin.multiplier} rounds of \code{burnin} iterations will be run.
After each round, a burn-in chain will be save for checkpoint.}

\item{burnin.checkpoint}{Default is False. If True, a checkpoint for burnin will be created.}
}
\value{
A list with 2 elements: \describe{
\item{\code{hdplist}}{A list representation of
an \code{\link[hdpx]{hdpState-class}} object.}
\item{likelihood}{A numeric vector with the likelihood at each iteration.}
}
}
\description{
Prepare an \code{\link[hdpx]{hdpState-class}} object and run the Gibbs sampling burnin.
}
