% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChainBurnin.R
\name{ChainBurnin}
\alias{ChainBurnin}
\title{Prepare an \code{\link[hdpx]{hdpState-class}} object and run the Gibbs sampling burnin.}
\usage{
ChainBurnin(
  hdp.state,
  seedNumber = 1,
  burnin = 5000,
  cpiter = 3,
  burnin.verbosity = 0,
  burnin.multiplier = 2,
  checkpoint = TRUE
)
}
\arguments{
\item{hdp.state}{An \code{\link[hdpx]{hdpState-class}} object or a list
representation of an \code{\link[hdpx]{hdpState-class}} object.}

\item{seedNumber}{Used to create the checkpoint file name if \code{checkpoint}
is \code{TRUE}.}

\item{burnin}{\code{burnin}. The number of burn-in iterations in
one batch. The total number of burnin iterations is
\code{burnin * burnin.multiplier}.
Passed to argument burnin of hdpx::hdp_burnin}

\item{cpiter}{The number of iterations of concentration
parameter sampling
to perform after each main Gibbs-sample iteration. (See Teh et al.,
"Hierarchical Dirichlet Processes", Journal of the American Statistical
Association 2006;101(476):1566-1581
(https://doi.org/10.1198/016214506000000302).)
in package hdpx.}

\item{burnin.verbosity}{Pass to \code{\link[hdpx]{hdp_burnin}}
\code{verbosity}. Verbosity of debugging statements.
#'}

\item{burnin.multiplier}{Run \code{burnin.multiplier} rounds of \code{burnin}
iterations.
If \code{checkpoint} is \code{TRUE}, save the burnin chain (see
parameter \code{checkpoint}.)
The diagnostic plot \code{diagnostics.likelihood.pdf} can help
determine if the chain is stationary.
The burnin can be continued from a checkpoint file
with \code{\link{ExtendBurnin}} (see argument \code{checkpoint}).}

\item{checkpoint}{If \code{TRUE}, a checkpoint
file called mSigHdp.burnin.checkpoint.\emph{seed_number}.Rdata.
will be created
in the current working directory.}
}
\value{
A list with 2 elements: \describe{
\item{\code{hdplist}}{A list representation of
an \code{\link[hdpx]{hdpState-class}} object.}
\item{likelihood}{A numeric vector with the likelihood at each iteration.}
}
}
\description{
Prepare an \code{\link[hdpx]{hdpState-class}} object and run the Gibbs sampling burnin.
}
\keyword{internal}
