% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineChainsAndExtractSigs.R
\name{CombineChainsAndExtractSigs}
\alias{CombineChainsAndExtractSigs}
\title{Extract components and exposures from multiple posterior sample chains
This function returns signatures with high confidence (found in more than 90\% #' posterior samples)}
\usage{
CombineChainsAndExtractSigs(
  clean.chlist,
  input.catalog,
  multi.types,
  verbose = TRUE,
  cos.merge = 0.9,
  confident.prop = 0.9,
  noise.prop = 0.1,
  hc.cutoff = 0.1
)
}
\arguments{
\item{clean.chlist}{A list of \code{\link[hdpx]{hdpSampleChain-class}} objects.
Each element is the result of one posterior sample chain.}

\item{input.catalog}{Input spectra catalog as a matrix or
in \code{\link[ICAMS]{ICAMS}} format.}

\item{multi.types}{A logical scalar or
a character vector.
If \code{FALSE}, The HDP analysis
will regard all input spectra as one tumor type.
HDP structure as one parent node for all tumors

If \code{TRUE}, the HDP analysis
will infer tumor types based on the string before "::" in their names.
e.g. tumor type for "SA.Syn.Ovary-AdenoCA::S.500" would be "SA.Syn.Ovary-AdenoCA"
HDP structure as a grandparent node for whole data and
one parent node for each tumor type

If \code{multi.types} is a character vector, then it should be of the same length
as the number of columns in \code{input.catalog}, and each value is the
name of the tumor type of the corresponding column in \code{input.catalog}.

e.g. \code{c("SA.Syn.Ovary-AdenoCA", "SA.Syn.Kidney-RCC")}.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}

\item{cos.merge}{The cosine similarity threshold for merging raw clusters
from the posterior sampling chains into "components" i.e. signatures;
passed to \code{\link[hdpx]{extract_components_from_clusters}}.}

\item{confident.prop}{Pass to \code{\link[hdpx]{interpret_components}}.
clusters with at least \code{confident.prop} of posterior
samples are high confident signatures}

\item{noise.prop}{Pass to \code{\link[hdpx]{interpret_components}}.
Clusters with less than \code{noise.prop} of posterior samples
are noise signatures}

\item{hc.cutoff}{Pass to \code{\link[hdpx]{extract_components_from_clusters}}. The cutoff of
height of hierarchical clustering dendrogram}
}
\value{
Invisibly, a list with the following elements:\describe{
\item{signature}{The extracted signature profiles as a matrix;
rows are mutation types, columns are signatures including high confident signatures -'hdp' signatures
and moderate confident signatures - 'potential hdp' signatures.}

\item{signature.post.samp.number}{A dataframe with two columns. The first column corresponds to each signature
in \code{signature} and the second columns contains the number of posterior
samples that found the raw clusters contributing to the signature.}

\item{signature.cdc}{A \code{\link[hdpx]{comp_dp_counts}} like dataframe. Each column corresponds to the sum of
all \code{\link[hdpx]{comp_dp_counts}} matrices of the raw clusters contributing to each
signature in code{signature}}

\item{exposureProbs}{The inferred exposures as a matrix of mutation probabilities;
rows are signatures, columns are samples (e.g. tumors).}

\item{noise.signature}{The extracted signature profiles as a matrix; rows are mutation types,
columns are signatures with less than \code{noise.prop} of posterior samples.}

\item{noise.post.samp.number}{A data frame with two columns. The first column corresponds to each signature
in \code{noise.signature} and the second columns contains the number of posterior
samples that found the raw clusters contributing to the signature.}

\item{noise.cdc}{A \code{\link[hdpx]{comp_dp_counts}} like data frame. Each column corresponds to the sum of
all \code{\link[hdpx]{comp_dp_counts}} matrices of the raw clusters contributing to each
signature in code{noise.signature}}

\item{extracted.retval}{A list object returned from code{\link[hdpx]{interpret_components}}.}
}
}
\description{
Extract components and exposures from multiple posterior sample chains
This function returns signatures with high confidence (found in more than 90\% #' posterior samples)
}
