% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineChainsAndExtractSigs.R
\name{CombineChainsAndExtractSigs}
\alias{CombineChainsAndExtractSigs}
\title{Extract components and exposures from multiple posterior sample chains
This function returns signatures with high confidence (found in more than 90\% #' posterior samples)}
\usage{
CombineChainsAndExtractSigs(
  clean.chlist,
  input.catalog,
  verbose = TRUE,
  high.confidence.prop = 0.9,
  hc.cutoff = 0.1
)
}
\arguments{
\item{clean.chlist}{A list of \code{\link[hdpx]{hdpSampleChain-class}} objects.
Each element is the result of one posterior sample chain.}

\item{input.catalog}{Input spectra catalog as a matrix or
in \code{\link[ICAMS]{ICAMS}} format.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}

\item{high.confidence.prop}{Pass to \code{\link[hdpx]{interpret_components}}.
raw clusters (mutation cluster)
found in \eqn{>=} \code{high.confidence.prop} proportion of posterior
samples are signatures with high confidence.}

\item{hc.cutoff}{Pass to \code{\link[hdpx]{extract_components_from_clusters}}. The cutoff of
height of hierarchical clustering dendrogram, used in combining
raw clusters (mutation clusters) into agreggated clusters.}
}
\value{
Invisibly, a list with the following elements:\describe{
\item{signature}{The extracted signature profiles as a matrix;
rows are mutation types, columns are signatures with
high confidence.}

\item{signature.post.samp.number}{A data frame with two columns. The first
column corresponds to each signature in \code{signature}
and the second columns contains the number of posterior
samples that found the raw clusters contributing to the signature.}

\item{signature.cdc}{A numeric data frame. Each column corresponds
to the sum of all mutations contributing to each signature in \code{signature}}

\item{exposureProbs}{The inferred exposures as a matrix of mutation probabilities;
rows are signatures, columns are samples (e.g. tumors). This is
similar to \code{signature.cdc} but every column was normalized to sum of 1}

\item{low.confidence.signature}{The profiles of signatures extracted with low confidence as a matrix; rows are mutation types,
columns are signatures with less than \code{high.confidence.prop} of posterior samples }

\item{low.confidence.post.samp.number}{A data frame with two columns. The first column corresponds
to each signature in \code{low.confidence.signature} and the second
column contains the number of posterior samples that found
the raw clusters contributing to the signature.}

\item{low.confidence.cdc}{A numeric data frame. Each column corresponds
to the sum of all mutations contributing to each signature in \code{low.confidence.signature}}

\item{extracted.retval}{A list object returned from code{\link[hdpx]{extract_components_from_clusters}}.}

}
}
\description{
Extract components and exposures from multiple posterior sample chains
This function returns signatures with high confidence (found in more than 90\% #' posterior samples)
}
