% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineChainsAndExtractSigs.R
\name{CombineChainsAndExtractSigs}
\alias{CombineChainsAndExtractSigs}
\title{Extract signatures etc. from multiple Gibbs sample chains}
\usage{
CombineChainsAndExtractSigs(
  clean.chlist,
  input.catalog,
  verbose = FALSE,
  high.confidence.prop = 0.9,
  hc.cutoff = 0.1
)
}
\arguments{
\item{clean.chlist}{A list of \code{\link[hdpx]{hdpSampleChain-class}}
objects (from package hdpx), typically returned from
\code{ParallelGibbsSample}.
Each element must be the result of one posterior sample chain.}

\item{input.catalog}{Input spectra catalog as a matrix.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}

\item{high.confidence.prop}{Raw clusters of mutations
found in \eqn{>=} \code{high.confidence.prop} proportion of posterior
samples are signatures with high confidence.}

\item{hc.cutoff}{The cutoff of
height of the hierarchical clustering dendrogram used in combining
raw clusters of mutations into aggregated clusters.}
}
\value{
Invisibly, a list with the following elements:\describe{
\item{signature}{The extracted signature profiles as a matrix;
rows are mutation types, columns are signatures with
high confidence.}

\item{signature.post.samp.number}{A data frame with two columns. The first
column corresponds to each signature in \code{signature}
and the second columns contains the number of posterior
samples that found the raw clusters contributing to the signature.}

\item{signature.cdc}{
A numeric data frame. Columns correspond to signatures
as in \code{signature}. Rows correspond to either biological
samples or to parent and grandparent Dirichlet processes.}

\item{exposureProbs}{The inferred exposures as a matrix
of mutation probabilities;
rows are signatures, columns are samples (e.g. tumors). This is
similar to \code{signature.cdc}, but every column was normalized
to sum to 1.}

\item{low.confidence.signature}{The profiles of signatures extracted
with low confidence as a matrix; rows are mutation types,
columns are signatures with <
\code{high.confidence.prop} of posterior samples.}

\item{low.confidence.post.samp.number}{
Analogous to \code{signature.post.samp.number}, except that
this one is for
signatures in \code{low.confidence.signature}.}

\item{low.confidence.cdc}{Analogous to
\code{signature.cdc}, except that columns in this
matrix correspond to columns
in \code{low.confidence.signature}.}

\item{extracted.retval}{A list object returned from
\code{\link[hdpx]{extract_components_from_clusters}}
in package hdpx.}

}
}
\description{
Extract signatures etc. from multiple Gibbs sample chains
}
