% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineChainsAndExtractSigs.R
\name{CombineChainsAndExtractSigs}
\alias{CombineChainsAndExtractSigs}
\title{Extract components and exposures from multiple posterior sample chains
This function returns signatures with high confidence (found in more than 90\% #' posterior samples)}
\usage{
CombineChainsAndExtractSigs(
  clean.chlist,
  input.catalog,
  multi.types,
  verbose = TRUE,
  cos.merge = 0.9,
  confident.prop = 0.9,
  noise.prop = 0.1
)
}
\arguments{
\item{clean.chlist}{A list of \code{\link[hdpx]{hdpSampleChain-class}} objects.
Each element is the result of one posterior sample chain.}

\item{input.catalog}{Input spectra catalog as a matrix or
in \code{\link[ICAMS]{ICAMS}} format.}

\item{multi.types}{A logical scalar or
a character vector.
If \code{FALSE}, The HDP analysis
will regard all input spectra as one tumor type.

If \code{TRUE}, the HDP analysis
will infer tumor types based on the string before "::" in their names.
e.g. tumor type for "SA.Syn.Ovary-AdenoCA::S.500" would be "SA.Syn.Ovary-AdenoCA"

If \code{multi.types} is a character vector, then it should be of the same length
as the number of columns in \code{input.catalog}, and each value is the
name of the tumor type of the corresponding column in \code{input.catalog}.

e.g. \code{c("SA.Syn.Ovary-AdenoCA", "SA.Syn.Kidney-RCC")}.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}

\item{cos.merge}{The cosine similarity threshold for merging raw clusters
from the posterior sampling chains into "components" i.e. signatures;
passed to \code{\link[hdpx]{extract_sigs_from_clusters}}.}

\item{confident.prop}{passed to \code{\link[hdpx]{extract_sigs_from_clusters}}}

\item{noise.prop}{passed to \code{\link[hdpx]{extract_sigs_from_clusters}}}
}
\value{
Invisibly, a list with the following elements:\describe{
\item{signature}{The extracted signature profiles as a matrix;
rows are mutation types, columns are
samples (e.g. tumors).}

\item{exposure}{The inferred exposures as a matrix of mutation counts;
rows are signatures, columns are samples (e.g. tumors).}

\item{multi.chains}{A \code{\link[hdpx]{hdpSampleMulti-class}} object.
This object has the method \code{\link[hdpx]{chains}} which returns
a list of \code{\link[hdpx]{hdpSampleChain-class}} objects. Each of these
sample chains objects has a method \code{\link[hdpx]{final_hdpState}}
(actually the methods seems to be just \code{hdp})
that returns the \code{hdpState} from which it was generated.}

#'
}
}
\description{
Extract components and exposures from multiple posterior sample chains
This function returns signatures with high confidence (found in more than 90\% #' posterior samples)
}
