% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombinePosteriorChains.R
\name{CombinePosteriorChains}
\alias{CombinePosteriorChains}
\title{Extract components and exposures from multiple posterior sample chains}
\usage{
CombinePosteriorChains(
  clean.chlist,
  input.catalog,
  multi.types,
  cluster.method = "kmedians",
  verbose = TRUE,
  cos.merge = 0.9,
  categ.CI = 0.95,
  exposure.CI = 0.95,
  min.sample = 1
)
}
\arguments{
\item{clean.chlist}{A list of \code{\link[hdpx]{hdpSampleChain-class}} objects.
Each element is the result of one posterior sample chain.}

\item{input.catalog}{Input spectra catalog as a matrix or
in \code{\link[ICAMS]{ICAMS}} format.}

\item{multi.types}{A logical scalar or
a character vector.
If \code{FALSE}, The HDP analysis
will regard all input spectra as one tumor type.

If \code{TRUE}, the HDP analysis
will infer tumor types based on the string before "::" in their names.
e.g. tumor type for "SA.Syn.Ovary-AdenoCA::S.500" would be "SA.Syn.Ovary-AdenoCA"

If \code{multi.types} is a character vector, then it should be of the same length
as the number of columns in \code{input.catalog}, and each value is the
name of the tumor type of the corresponding column in \code{input.catalog}.

e.g. \code{c("SA.Syn.Ovary-AdenoCA", "SA.Syn.Kidney-RCC")}.}

\item{cluster.method}{A \code{\link[flexclust]{kccaFamily}} object.
Deprecated. Future code will use \code{"kmedians"}.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}

\item{cos.merge}{The cosine similarity threshold for merging raw clusters
from the posterior sampling chains into "components" i.e. signatures;
passed to \code{\link[hdpx]{hdp_extract_components}}.}

\item{categ.CI}{A number the range \eqn{[0,1]}. The level of the confidence
interval used in step 4 of \code{\link{hdp_merge_and_extract_components}}.
This governs when "averaged raw cluster" get assigned to component 0,
i.e. if the the confidence interval overlaps 0. Lower values
make it less likely that an averaged raw cluster will be assigned to
component 0. The CI in question is for the number of mutations in
a given mutation class (e.g. ACA > AAA, internally called a
"category"). If, for every mutation class, this CI overlaps 0,
then the averaged raw cluster goes to component 0.}

\item{exposure.CI}{A number in the range \eqn{[0,1]}. The level of
the confidence interval used in step 5 of hdp_merge_and_extract_components.
The CI in question here for the total number of
mutations assigned to an averaged raw cluster.}

\item{min.sample}{A "component" (i.e. signature) must have at least
this many samples; passed to \code{\link[hdpx]{hdp_merge_and_extract_components}}.}
}
\value{
Invisibly, a list with the following elements:\describe{

\item{signature}{The extracted signature profiles as a matrix;
rows are mutation types, columns are
samples (e.g. tumors).}

\item{exposure}{The inferred exposures as a matrix of mutation counts;
rows are signatures, columns are samples (e.g. tumors).}

\item{multi.chains}{A \code{\link[hdpx]{hdpSampleMulti-class}} object.
This object has the method \code{\link[hdpx]{chains}} which returns
a list of \code{\link[hdpx]{hdpSampleChain-class}} objects. Each of these
sample chains objects has a method \code{\link[hdpx]{final_hdpState}}
(actually the methods seems to be just \code{hdp})
that returns the \code{hdpState} from which it was generated.}

\item{sum_raw_clusters_after_cos_merge}{A matrix containing aggregated spectra of raw clusters after cosine
similarity merge step in \code{\link[hdpx]{hdp_merge_and_extract_components}}.}

\item{sum_raw_clusters_after_nonzero_categ}{A matrix containing aggregated spectra of raw clusters after non-zero category selecting
step in \code{\link[hdpx]{hdp_merge_and_extract_components}}.}

\item{clust_hdp0_ccc4}{A matrix containing aggregated spectra of raw clusters moving to hdp.0 after non-zero category selection step in              \code{\link[hdpx]{hdp_merge_and_extract_components}}.}

\item{clust_hdp0_ccc5}{A matrix containing aggregated spectra of raw clusters moving to hdp.0 after non-zero observation selection step in           \code{\link[hdpx]{hdp_merge_and_extract_components}}.}\preformatted{  \}
}
}
\description{
Extract components and exposures from multiple posterior sample chains
}
