% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComponentDiagnosticPlotting.R
\name{ComponentDiagnosticPlotting}
\alias{ComponentDiagnosticPlotting}
\title{Generate diagnostic plots from the return value of \code{\link{CombineChainsAndExtractSigs}}}
\usage{
ComponentDiagnosticPlotting(retval, input.catalog, out.dir, verbose)
}
\arguments{
\item{retval}{Return from \code{\link{CombineChainsAndExtractSigs}.}}

\item{input.catalog}{Input spectra catalog as a matrix.}

\item{out.dir}{If not \code{NULL} then
a character string specifying
a directory that will be created for the output, including
csv files and plots (pdfs) of extracted signatures and their exposures.
If \code{NULL} no directory will be created and no files
will be generated.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}
}
\description{
Generate diagnostic plots from the return value of \code{\link{CombineChainsAndExtractSigs}}
}
\details{
Generates the plots
\itemize{
\item \code{diagnostics.signature.exposure.each.sample.pdf}
\item \code{diagnostics.components.in.which.gibbs.samples.pdf}
\item \code{diagnostics.likelihood.pdf}
\item \code{diagnostics.numcluster.pdf}
\item \code{diagnostics.signatures.pdf}
}
}
\keyword{internal}
