% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComponentDiagnosticPlotting.R
\name{ComponentDiagnosticPlotting}
\alias{ComponentDiagnosticPlotting}
\title{Diagnostic plot for a hdpSampleMulti object}
\usage{
ComponentDiagnosticPlotting(retval, input.catalog, out.dir, verbose)
}
\arguments{
\item{retval}{output from \code{CombinePosteriorChains.
}}

\item{input.catalog}{ground truth catalog}

\item{out.dir}{Directory that will be created for the output;
if \code{overwrite} is \code{FALSE} then
abort if \code{out.dir} already exits.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}
}
\description{
Diagnostic plot for a hdpSampleMulti object
}
