% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComponentDiagnosticPlotting.R
\name{ComponentDiagnosticPlotting}
\alias{ComponentDiagnosticPlotting}
\title{Generate multiple plots for for a hdpSampleMulti object.}
\usage{
ComponentDiagnosticPlotting(
  retval,
  input.catalog,
  out.dir,
  verbose,
  IS.ICAMS = T
)
}
\arguments{
\item{retval}{Return from \code{\link{CombineChainsAndExtractSigs}}}

\item{input.catalog}{Input spectra catalog as a matrix or
in \code{\link[ICAMS]{ICAMS}} format.}

\item{out.dir}{Directory that will be created for the output;
if \code{overwrite} is \code{FALSE} then
abort if \code{out.dir} already exits.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}

\item{IS.ICAMS}{If TRUE, then plot diagnostics.hdp.signature.exposure.each.sample.pdf.}
}
\description{
Generate multiple plots for for a hdpSampleMulti object.
}
\details{
Generates the plots diagnostics.hdp.signature.exposure.each.sample.pdf,
diagnostics.component.distribution.in.posterior.samples.pdf,
diagnostics.likelihood.pdf,
diagnostics.numcluster.pdf,
diagnostics.signatures.pdf
}
