% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComponentDiagnosticPlotting.R
\name{ComponentDiagnosticPlotting}
\alias{ComponentDiagnosticPlotting}
\title{Generate multiple plots from the return value of \code{\link{CombineChainsAndExtractSigs}}.}
\usage{
ComponentDiagnosticPlotting(retval, input.catalog, out.dir, verbose)
}
\arguments{
\item{retval}{Return from \code{\link{CombineChainsAndExtractSigs}.}}

\item{input.catalog}{Input spectra catalog as a matrix.}

\item{out.dir}{Directory that will be created for the output, including
csv files and plots (pdfs) of extracted signatures and their exposures.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}
}
\description{
Generate multiple plots from the return value of \code{\link{CombineChainsAndExtractSigs}}.
}
\details{
Generates the plots
\itemize{
\item \code{diagnostics.hdp.signature.exposure.each.sample.pdf}
\item \code{diagnostics.component.distribution.in.posterior.samples.pdf}
\item \code{diagnostics.likelihood.pdf}
\item \code{diagnostics.numcluster.pdf}
\item \code{diagnostics.signatures.pdf}
}
}
