% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample}
\alias{downsample}
\title{Down sample a vector of integers; nonsensical for negative values}
\usage{
downsample(x, thres = 3000)
}
\arguments{
\item{x}{Vector of integers}

\item{thres}{Values $le$ \code{thres} are unmodified; for values of
\code{thres} > 3,000 some of the return values will also not be
reduced.}
}
\value{
A vector of integers (type \code{numeric}) of the same
length as \code{x}, with elements that were $le$ \code{thres} in
\code{x} reduced.
}
\description{
Down sample a vector of integers; nonsensical for negative values
}
