% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample}
\alias{downsample}
\title{Down sample a vector of type \code{numeric}; nonsensical for negative values.}
\usage{
downsample(x, thres = 3000)
}
\arguments{
\item{x}{A \code{numeric} vector.}

\item{thres}{If \code{thres} >= 3,000, then in
all input values > \code{thres} are
downsampled.
If \code{thres} < 3,000,
some input values > \code{thres} are
downsampled.}
}
\value{
A vector of integers (type \code{numeric}) of the same
length as \code{x}, downsampled as described in the
documentation for the \code{thres} argument.
}
\description{
Down sample a vector of type \code{numeric}; nonsensical for negative values.
}
