% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample_spectra}
\alias{downsample_spectra}
\title{Down sample a set of mutational spectra}
\usage{
downsample_spectra(spec, thres)
}
\arguments{
\item{spec}{Input spectra as a numerical matrix or similar \code{data.frame};
each column is a spectrum, each row is a mutation type (e.g. CAG -> CTG).}

\item{thres}{See \code{\link{downsample}}.}
}
\value{
A numeric matrix with same shape as \code{spec},
with the entries each column reduced based on
the ratio of
\code{\link{downsample}(colSums(spec))} to
\code{colSums(spec)}.
}
\description{
Down sample a set of mutational spectra
}
