% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample_spectra}
\alias{downsample_spectra}
\title{Downsample a set of mutational spectra}
\usage{
downsample_spectra(spec, downsample_threshold = NULL, thres = NULL)
}
\arguments{
\item{spec}{Input spectra as a numerical matrix or similar \code{data.frame};
each column is a spectrum, each row is a mutation type (e.g. CAG -> CTG).}

\item{downsample_threshold}{See \code{\link{downsample}}
and \code{\link{show_downsample_curves}}.}

\item{thres}{Deprecated alternative to \code{downsample_threshold}.}
}
\value{
A list with elements:
\itemize{
\item \code{down_spec}: A numeric matrix with same shape as \code{spec},
with the mutation counts in each column reduced based on
the corresponding ratio in the second element of this list, \code{down_factor}.
\item \code{down_factor} Numeric vector of the ratios of
\code{\link{downsample}(colSums(spec))} to
\code{colSums(spec)}.
}
}
\description{
Downsample a set of mutational spectra
}
