% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtendBurnin.R
\name{ExtendBurnin}
\alias{ExtendBurnin}
\title{Extend Burn in iteration for a list representation of
an \code{\link[hdpx]{hdpState-class}} object. This list
is an output from \code{\link[hdpx]{hdp_burnin}} or
\code{ActivateandBurnin}.}
\usage{
ExtendBurnin(hdplist, seedNumber = 1, burnin = 4000, cpiter = 3, verbosity = 0)
}
\arguments{
\item{hdplist}{A list representation of
an \code{\link[hdpx]{hdpState-class}} object}

\item{seedNumber}{A random seed for setting the environment of
\code{\link[hdpx]{hdp_burnin}}.}

\item{burnin}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{burnin}.}

\item{cpiter}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{cpiter}.}

\item{verbosity}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{verbosity}.}
}
\value{
A list with hdp object after burn-in iteration and likelihood of iteration
}
\description{
Extend Burn in iteration for a list representation of
an \code{\link[hdpx]{hdpState-class}} object. This list
is an output from \code{\link[hdpx]{hdp_burnin}} or
\code{ActivateandBurnin}.
}
