% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtendBurnin.R
\name{ExtendBurnin}
\alias{ExtendBurnin}
\title{Extend burnin iterations generated from \code{\link{Burnin}}}
\usage{
ExtendBurnin(
  previous.burnin.output,
  burnin = 5000,
  cpiter = 3,
  burnin.verbosity = 0,
  seedNumber = NULL
)
}
\arguments{
\item{previous.burnin.output}{Output from
\code{\link{Burnin}} or the file path of
a checkpoint file written by \code{\link{Burnin}}.}

\item{burnin}{The number of burnin iterations to perform.}

\item{cpiter}{The number of iterations of concentration
parameter sampling
to perform after each main Gibbs-sample iteration. (See Teh et al.,
"Hierarchical Dirichlet Processes", Journal of the American Statistical
Association 2006;101(476):1566-1581
(https://doi.org/10.1198/016214506000000302).)}

\item{burnin.verbosity}{Number that controls whether progress
messages are printed.}

\item{seedNumber}{A random seed for reproducible results.}
}
\value{
The same type of object as returned from
\code{\link{Burnin}.}

The envisioned application is extending burnins from burnin checkpoints.
}
\description{
Extend burnin iterations generated from \code{\link{Burnin}}
}
