% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generateppindex.R
\name{Generateppindex}
\alias{Generateppindex}
\title{Generate index for a HDP structure and num.tumor.types for other functions}
\usage{
Generateppindex(multi.types, input.catalog)
}
\arguments{
\item{multi.types}{A logical scalar or
a character vector.

If \code{FALSE}, The HDP analysis
will regard all input spectra as one tumor type, and the
HDP structure will have one parent node for all tumors.

If \code{TRUE}, Sample IDs in \code{input.catalog} must have the
form \emph{sample_type}::\emph{sample_id}.

If a character vector, then its length must be \code{ncol(input.catalog}),
and each value is the sample type of the corresponding column in \code{input.catalog},
e.g. \code{c(rep("Type-A", 23), rep("Type-B", 10))} for 23 Type-A samples and 10 Type-B
samples.

If not \code{FALSE}, HDP will have one parent node for each sample type
and one grandparent node.}

\item{input.catalog}{Input spectra catalog as a matrix or
in \code{\link[ICAMS]{ICAMS}} format.}
}
\description{
Generate index for a HDP structure and num.tumor.types for other functions
}
\keyword{internal}
