% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generateppindex.R
\name{Generateppindex}
\alias{Generateppindex}
\title{Generate index for a HDP structure}
\usage{
Generateppindex(multi.types, one.parent.hack, input.catalog)
}
\arguments{
\item{multi.types}{A logical scalar or
a character vector.
If \code{FALSE}, The HDP analysis
will regard all input spectra as one tumor type.

If \code{TRUE}, the HDP analysis
will infer tumor types based on the string before "::" in their names.
e.g. tumor type for "SA.Syn.Ovary-AdenoCA::S.500" would be "SA.Syn.Ovary-AdenoCA"

If \code{multi.types} is a character vector, then it should be of the same length
as the number of columns in \code{input.catalog}, and each value is the
name of the tumor type of the corresponding column in \code{input.catalog}.

e.g. \code{c("SA.Syn.Ovary-AdenoCA", "SA.Syn.Kidney-RCC")}.}

\item{one.parent.hack}{IF TRUE, the vector of parents for the Dirichlet
processes looks like c(0, 1, 1, 1, 1, ...), not c(0, 1, 2, 2, 2, ....).}

\item{input.catalog}{Input spectra catalog as a matrix or
in \code{\link[ICAMS]{ICAMS}} format.}
}
\description{
Generate index for a HDP structure
}
