% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsSamplingAfterBurnin.R
\name{GibbsSamplingAfterBurnin}
\alias{GibbsSamplingAfterBurnin}
\title{Start Gibbs sampling after burnin}
\usage{
GibbsSamplingAfterBurnin(
  burnin.output = burnin.output,
  post.n = post.n,
  post.space = post.space,
  post.cpiter = 3,
  post.verbosity = 0,
  seedNumber = NULL
)
}
\arguments{
\item{burnin.output}{A path to burnin checkpoint Rdata or An S4 object from \code{ChainBurnin}}

\item{post.n}{Pass to \code{\link[hdpx]{hdp_posterior_sample}}
\code{n}.The number of posterior samples to collect.}

\item{post.space}{Pass to \code{\link[hdpx]{hdp_posterior_sample}}
\code{space}. The number of iterations between collected samples.}

\item{post.cpiter}{Pass to \code{\link[hdpx]{hdp_posterior_sample}} and
\code{\link[hdpx]{hdp_burnin}} \code{cpiter}.The number of iterations of concentration
parameter sampling to perform after each iteration}

\item{post.verbosity}{Pass to \code{\link[hdpx]{hdp_posterior_sample}}
\code{verbosity}. Verbosity of debugging statements.
No need to change unless for development purpose. Default is 0.}
}
\value{
A hdpSampleChain object with the salient information from each
posterior sample. See \code{\link{hdpSampleChain-class}}
}
\description{
Start Gibbs sampling after burnin
}
\keyword{internal}
