% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsSamplingAfterBurnin.R
\name{GibbsSamplingAfterBurnin}
\alias{GibbsSamplingAfterBurnin}
\title{Start Gibbs sampling on one chain after burnin.}
\usage{
GibbsSamplingAfterBurnin(
  burnin.output = burnin.output,
  post.n = post.n,
  post.space = post.space,
  post.cpiter = 3,
  post.verbosity = 0,
  seedNumber = NULL
)
}
\arguments{
\item{burnin.output}{A path to burnin checkpoint Rdata
or to an S4 object from \code{ChainBurnin}}

\item{post.n}{The number of posterior samples to collect.}

\item{post.space}{The number of iterations between collected samples.}

\item{post.cpiter}{The number of iterations of concentration
parameter sampling to perform after each main
Gibbs sampling iteration.}

\item{post.verbosity}{Verbosity of debugging statements.
No need to change unless for testing or debugging. Default is 0.}

\item{seedNumber}{A random seed that ensures ensures reproducible
results.}
}
\value{
An \code{hdpSampleChain}
object with the salient information from each
posterior sample. See \code{\link{hdpSampleChain-class}}.
}
\description{
This function might be used to start Gibbs sampling after
\code{\link{ExtendBurnin}}.
}
