% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSamplesHighSigExp.R
\name{PlotSamplesHighSigExp}
\alias{PlotSamplesHighSigExp}
\title{Plot hdp signature exposure in each sample.
This function returns the plot of
top 5 samples with the highest exposure to a signature.
Each spectrum's title is in the form of: \emph{Sample Name
(Proportion of Signature Assignment)}.
This function is here because it is
specific for the signature extraction application.}
\usage{
PlotSamplesHighSigExp(
  retval,
  hdpsample,
  input.catalog,
  col_comp = NULL,
  incl_numdata_plot = F,
  ylab_numdata = "Number of data items",
  ylab_exp = "Component exposure",
  leg.title = "Component",
  cex.names = 0.6,
  cex.axis = 0.7,
  mar = c(4, 4, 2, 0.5),
  oma = c(1.5, 1.5, 1, 1)
)
}
\arguments{
\item{retval}{An object return from \code{\link[hdpx]{extract_ccc_from_hdp}}
in package hdpx.}

\item{hdpsample}{A \code{\link[hdpx]{hdpSampleChain-class}} or
\code{\link[hdpx]{hdpSampleMulti-class}} object including output
from \code{\link[hdpx]{extract_components}} in package hdpx.}

\item{input.catalog}{Input spectra catalog as a matrix or
in \code{\link[ICAMS]{ICAMS}} format.}

\item{col_comp}{Colors of each component, from 0 to the max number.
If NULL, default colors will be used.}

\item{incl_numdata_plot}{Logical - should an upper barplot indicating the number of
data items per DP be included? (Default TRUE).}

\item{ylab_numdata}{Vertical axis label for \code{numdata} plot.}

\item{ylab_exp}{Vertical axis label for exposure plot.}

\item{leg.title}{Legend title.}

\item{cex.names}{Expansion factor for bar labels (\code{dpnames})
in exposure plot.}

\item{cex.axis}{Expansion factor for vertical-axis annotation.}

\item{mar}{See ?par}

\item{oma}{See ?par}
}
\description{
Plot hdp signature exposure in each sample.
This function returns the plot of
top 5 samples with the highest exposure to a signature.
Each spectrum's title is in the form of: \emph{Sample Name
(Proportion of Signature Assignment)}.
This function is here because it is
specific for the signature extraction application.
}
\keyword{internal}
