% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepInit.R
\name{PrepInit}
\alias{PrepInit}
\title{Initialize hdp object
Allocate process index for hdp initialization.
Prepare for hdp_init}
\usage{
PrepInit(multi.types, input.catalog, verbose, K.guess)
}
\arguments{
\item{multi.types}{TODO}

\item{input.catalog}{TODO}

\item{verbose}{TODO}

\item{K.guess}{TODO}
}
\description{
Initialize hdp object
Allocate process index for hdp initialization.
Prepare for hdp_init
}
