% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunHdpxParallel.R
\name{RunHdpxParallel}
\alias{RunHdpxParallel}
\title{Extract mutational signatures from a set of mutational spectra.}
\usage{
RunHdpxParallel(
  input.catalog,
  seedNumber = 123,
  K.guess,
  multi.types = TRUE,
  verbose = TRUE,
  burnin = 1000,
  burnin.multiplier = 10,
  post.n = 200,
  post.space = 100,
  post.cpiter = 3,
  post.verbosity = 0,
  CPU.cores = 20,
  num.child.process = 20,
  high.confidence.prop = 0.9,
  hc.cutoff = 0.1,
  overwrite = TRUE,
  out.dir = NULL,
  gamma.alpha = 1,
  gamma.beta = 20,
  checkpoint = TRUE,
  prior.sigs = NULL,
  prior.pseudoc = NULL
)
}
\arguments{
\item{input.catalog}{Input spectra catalog as a matrix or
in \code{\link[ICAMS]{ICAMS}} format.}

\item{seedNumber}{A random seed that ensures ensures reproducible
results.}

\item{K.guess}{Suggested initial value of the number of
clusters. Usually, the number of clusters is two times of the number
of extracted signatures. Passed to \code{\link[hdpx]{dp_activate}} as
\code{initcc}.}

\item{multi.types}{A logical scalar or
a character vector.

If \code{FALSE}, The HDP analysis
will regard all input spectra as one tumor type, and the
HDP structure will have one parent node for all tumors.

If \code{TRUE}, Sample IDs in \code{input.catalog} must have the
form \emph{sample_type}::\emph{sample_id}.

If a character vector, then its length must be \code{ncol(input.catalog}),
and each value is the sample type of the corresponding column in \code{input.catalog},
e.g. \code{c(rep("Type-A", 23), rep("Type-B", 10))} for 23 Type-A samples and 10 Type-B
samples.

If not \code{FALSE}, HDP will have one parent node for each sample type
and one grandparent node.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}

\item{burnin}{Pass to \code{\link[hdpx]{hdp_burnin}}
\code{burnin}. The number of burn-in iterations}

\item{burnin.multiplier}{Run \code{burnin.multiplier} rounds of \code{burnin}
iterations.
If \code{checkpoint} is \code{TRUE}, save the burnin chain (see
parameter \code{checkpoint}.)
The diagnostic plot \code{diagnostics.likelihood.pdf} can help
determine if the chain is stationary.
The burnin can be continued from a checkpoint file
with \code{\link{ExtendBurnin}} (see argument \code{checkpoint}).}

\item{post.n}{Pass to \code{\link[hdpx]{hdp_posterior_sample}}
\code{n}.The number of posterior samples to collect.}

\item{post.space}{Pass to \code{\link[hdpx]{hdp_posterior_sample}}
\code{space}. The number of iterations between collected samples.}

\item{post.cpiter}{Pass to \code{\link[hdpx]{hdp_posterior_sample}} and
\code{\link[hdpx]{hdp_burnin}} \code{cpiter}.The number of iterations of concentration
parameter sampling to perform after each iteration}

\item{post.verbosity}{Pass to \code{\link[hdpx]{hdp_posterior_sample}}
\code{verbosity}. Verbosity of debugging statements.
No need to change unless for development purpose}

\item{CPU.cores}{Number of CPUs to use; this should be no more than \code{num.child.process}.}

\item{num.child.process}{Number of posterior sampling chains; can set to
1 for testing. We recommend 20 for real data analysis}

\item{high.confidence.prop}{Raw clusters of mutations
found in \eqn{>=} \code{high.confidence.prop} proportion of posterior
samples are signatures with high confidence.
(Passed to \code{\link[hdpx]{interpret_components}}).}

\item{hc.cutoff}{The cutoff of
height of the hierarchical clustering dendrogram used in combining
raw clusters of mutations into aggregated clusters.
(Passed to \code{\link[hdpx]{extract_components_from_clusters}}.)}

\item{overwrite}{If \code{TRUE} overwrite \code{out.dir}
if it exists, otherwise raise an error.}

\item{out.dir}{Directory that will be created for the output, including
csv files and plots (pdfs) of extracted signatures and their exposures.}

\item{gamma.alpha}{Shape parameter of
the gamma distribution prior for the Dirichlet process concentration
parameters (\eqn{\alpha_0} and all \eqn{\alpha_j} in
Figure B.1 of
https://www.repository.cam.ac.uk/bitstream/handle/1810/275454/Roberts-2018-PhD.pdf}

\item{gamma.beta}{Inverse scale parameter (rate parameter) of
the gamma distribution prior for the Dirichlet process concentration
parameters(\eqn{\beta_0} and all \eqn{\beta_j} in
Figure B.1 of
https://www.repository.cam.ac.uk/bitstream/handle/1810/275454/Roberts-2018-PhD.pdf

We recommend gamma.alpha = 1 and gamma.beta = 20 for
single-base-substitution signature extraction;
gamma.alpha = 1 and gamma.beta = 50  for doublet-base-substitution
and indel signature extraction}

\item{checkpoint}{If \code{TRUE}, then \itemize{
\item Checkpoint each final Gibbs sample
chain to the current working directory, in a file called
mSigHdp.sample.checkpoint.\emph{seed_number}.Rdata.
\item Periodically checkpoint the burnin state
to the current working directory, in files called
mSigHdp.burnin.checkpoint.\emph{seed_number}.Rdata.
}}

\item{prior.sigs}{DELETE ME LATER, NOT SUPPORTED. A matrix containing prior signatures.}

\item{prior.pseudoc}{DELETE ME LATER, NOT SUPPORTED. A numeric list. Pseudo counts of each prior signature. Recommended is 1000. In practice,
it may be advisable to put lower weights on prior signatures that you do not expect
to be present in your dataset, or even exclude some priors entirely.}
}
\value{
Invisibly, a list with the following elements:\describe{
\item{signature}{The extracted signature profiles as a matrix;
rows are mutation types, columns are signatures with
high confidence.}

\item{signature.post.samp.number}{A data frame with two columns. The first
column corresponds to each signature in \code{signature}
and the second columns contains the number of posterior
samples that found the raw clusters contributing to the signature.}

\item{signature.cdc}{A numeric data frame. Each column corresponds
to the sum of all mutations contributing to each signature in \code{signature}}

\item{exposureProbs}{The inferred exposures as a matrix of mutation probabilities;
rows are signatures, columns are samples (e.g. tumors). This is
similar to \code{signature.cdc} but every column was normalized to sum of 1}

\item{low.confidence.signature}{The profiles of signatures extracted with low confidence as a matrix; rows are mutation types,
columns are signatures with less than \code{high.confidence.prop} of posterior samples }

\item{low.confidence.post.samp.number}{A data frame with two columns. The first column corresponds
to each signature in \code{low.confidence.signature} and the second
column contains the number of posterior samples that found
the raw clusters contributing to the signature.}

\item{low.confidence.cdc}{A numeric data frame. Each column corresponds
to the sum of all mutations contributing to each
signature in \code{low.confidence.signature}}

\item{extracted.retval}{A list object returned from code{\link[hdpx]{extract_components_from_clusters}}.}

}
}
\description{
Extract mutational signatures from a set of mutational spectra.
}
