% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunHdpxParallel.R
\name{RunHdpxParallel}
\alias{RunHdpxParallel}
\title{Extract mutational signatures and optionally generate diagnostic plots to help
understand the results: e.g. the stability each extracted signature and
the tumors that drive the extraction of each signature.}
\usage{
RunHdpxParallel(
  input.catalog,
  seedNumber = 123,
  K.guess,
  multi.types = TRUE,
  verbose = TRUE,
  burnin = 1000,
  burnin.multiplier = 10,
  burnin.checkpoint = FALSE,
  post.n = 200,
  post.space = 100,
  post.cpiter = 3,
  post.verbosity = 0,
  CPU.cores = 20,
  num.child.process = 20,
  high.confidence.prop = 0.9,
  hc.cutoff = 0.1,
  overwrite = TRUE,
  out.dir = NULL,
  gamma.alpha = 1,
  gamma.beta = 20,
  gamma0.alpha = gamma.alpha,
  gamma0.beta = gamma.beta,
  checkpoint.chlist = TRUE,
  checkpoint.1.chain = TRUE,
  prior.sigs = NULL,
  prior.pseudoc = NULL,
  posterior.checkpoint = FALSE
)
}
\arguments{
\item{input.catalog}{Input spectra catalog as a matrix or
in \code{\link[ICAMS]{ICAMS}} format.}

\item{seedNumber}{A random seeds passed to \code{\link[hdpx]{dp_activate}}.}

\item{K.guess}{Suggested initial value of the number of
clusters. Usually, the number of clusters is two times of the number
of extracted signatures. Passed to \code{\link[hdpx]{dp_activate}} as
\code{initcc}.}

\item{multi.types}{A logical scalar or
a character vector.

If \code{FALSE}, The HDP analysis
will regard all input spectra as one tumor type, and the
HDP structure will have one parent node for all tumors.

If \code{TRUE}, Sample IDs in \code{input.catalog} must have the
form \emph{sample_type}::\emph{sample_id}.

If a character vector, then its length must be \code{ncol(input.catalog}),
and each value is the sample type of the corresponding column in \code{input.catalog},
e.g. \code{c(rep("Type-A", 23), rep("Type-B", 10))} for 23 Type-A samples and 10 Type-B
samples.

If not \code{FALSE}, HDP will have one parent node for each sample type
and one grandparent node.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}

\item{burnin}{Pass to \code{\link[hdpx]{hdp_burnin}}
\code{burnin}. The number of burn-in iterations}

\item{burnin.multiplier}{A checkpoint setting. \code{burnin.multiplier} rounds of \code{burnin} iterations will be run.
After each round, a burn-in chain will be save for checkpoint.
A total number of 10,000 iterations is recommended for most analysis. Therefore we set the default of \code{burnin} to
1000 and \code{burnin.multiplier} to 10. However, number of iterations can be adjusted based on the size of dataset.
The dataset with more mutations require longer burn-ins. According to our experience,
50,000 iterations are needed when analyzing all PCAWG7 genomes (2,780 samples).
The burnin can be continued from a checkpoint file with \code{\link{ExtendBurnin}}.}

\item{burnin.checkpoint}{If \code{TRUE}, a checkpoint for burn-in will be created.}

\item{post.n}{Pass to \code{\link[hdpx]{hdp_posterior_sample}}
\code{n}.The number of posterior samples to collect.}

\item{post.space}{Pass to \code{\link[hdpx]{hdp_posterior_sample}}
\code{space}. The number of iterations between collected samples.}

\item{post.cpiter}{Pass to \code{\link[hdpx]{hdp_posterior_sample}} and
\code{\link[hdpx]{hdp_burnin}} \code{cpiter}.The number of iterations of concentration
parameter sampling to perform after each iteration}

\item{post.verbosity}{Pass to \code{\link[hdpx]{hdp_posterior_sample}}
\code{verbosity}. Verbosity of debugging statements.
No need to change unless for development purpose}

\item{CPU.cores}{Number of CPUs to use; this should be no more than \code{num.child.process}.}

\item{num.child.process}{Number of posterior sampling chains; can set to
1 for testing. We recommend 20 for real data analysis}

\item{high.confidence.prop}{Pass to \code{\link[hdpx]{interpret_components}}.
raw clusters (mutation cluster)
found in \eqn{>=} \code{high.confidence.prop} proportion of posterior
samples are signatures with high confidence.}

\item{hc.cutoff}{Pass to \code{\link[hdpx]{extract_components_from_clusters}}. The cutoff of
height of hierarchical clustering dendrogram, used in combining
raw clusters (mutation clusters) into agreggated clusters.}

\item{overwrite}{If \code{TRUE} overwrite \code{out.dir} if it exists, otherwise raise an error.}

\item{out.dir}{Directory that will be created for the output;
if \code{overwrite} is \code{FALSE} then
abort if \code{out.dir} already exits.}

\item{gamma.alpha}{Shape parameter of
the gamma distribution prior for the Dirichlet process concentration
parameters; in this
function the gamma distributions for all Dirichlet processes,
except possibly the top level process, are the same.}

\item{gamma.beta}{Inverse scale parameter (rate parameter) of
the gamma distribution prior for the Dirichlet process concentration
parameters; in this
function the gamma distributions for all Dirichlet processes, except
possibly the top level process, are the same.

We recommend gamma.alpha = 1 and gamma.beta = 20 for single-base-substitution signatures extraction;
gamma.alpha = 1 and gamma.beta = 50  for doublet-base-substitution/INDEL signature extraction}

\item{gamma0.alpha}{See figure B.1 from Nicola Robert's thesis.
The shape parameter (\eqn{\alpha_0}) of the gamma
distribution priors for
the Dirichlet process concentration parameters  (\eqn{\gamma_0})
for \eqn{G_0}.}

\item{gamma0.beta}{See figure B.1 from Nicola Robert's thesis.
Inverse scale parameter (rate parameter, \eqn{\beta_0}) of the gamma
distribution priors for
the Dirichlet process concentration parameters (\eqn{\gamma_0})
for \eqn{G_0}.}

\item{checkpoint.chlist}{If \code{TRUE}, checkpoint the (unclean)
chlist to "initial.chlist.Rdata" in the current working directory.
and checkpoint the clean chlist to "clean.chlist.Rdata" in the
current working directory.}

\item{checkpoint.1.chain}{If \code{TRUE} checkpoint the sample
chain to current working directory, in a file called
sample.chain.\emph{seed_number}.Rdata.}

\item{prior.sigs}{A matrix containing prior signatures.}

\item{prior.pseudoc}{A numeric list. Pesudo counts of each prior signature. Recommended is 1000. In practice,
it may be advisable to put lower weights on prior signatures that you do not expect
to be present in your dataset, or even exclude some priors entirely.}

\item{posterior.checkpoint}{If \code{TRUE} checkpoint the posterior sampling after every 10
posterior samples collected}
}
\value{
Invisibly, a list with the following elements:\describe{
\item{signature}{The extracted signature profiles as a matrix;
rows are mutation types, columns are signatures with
high confidence.}

\item{signature.post.samp.number}{A data frame with two columns. The first
column corresponds to each signature in \code{signature}
and the second columns contains the number of posterior
samples that found the raw clusters contributing to the signature.}

\item{signature.cdc}{A numeric data frame. Each column corresponds
to the sum of all mutations contributing to each signature in \code{signature}}

\item{exposureProbs}{The inferred exposures as a matrix of mutation probabilities;
rows are signatures, columns are samples (e.g. tumors). This is
similar to \code{signature.cdc} but every column was normalized to sum of 1}

\item{low.confidence.signature}{The profiles of signatures extracted with low confidence as a matrix; rows are mutation types,
columns are signatures with less than \code{high.confidence.prop} of posterior samples }

\item{low.confidence.post.samp.number}{A data frame with two columns. The first column corresponds
to each signature in \code{low.confidence.signature} and the second
column contains the number of posterior samples that found
the raw clusters contributing to the signature.}

\item{low.confidence.cdc}{A numeric data frame. Each column corresponds
to the sum of all mutations contributing to each signature in \code{low.confidence.signature}}

\item{extracted.retval}{A list object returned from code{\link[hdpx]{extract_components_from_clusters}}.}

}
}
\description{
Extract mutational signatures and optionally generate diagnostic plots to help
understand the results: e.g. the stability each extracted signature and
the tumors that drive the extraction of each signature.
}
