% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetupAndActivate.R
\name{SetupAndActivate}
\alias{SetupAndActivate}
\title{Generate an HDP Gibbs sampling chain from a spectra catalog.}
\usage{
SetupAndActivate(
  input.catalog,
  seedNumber = 1,
  K.guess,
  multi.types = FALSE,
  verbose = TRUE,
  gamma.alpha = 1,
  gamma.beta = 20
)
}
\arguments{
\item{input.catalog}{Input spectra catalog as a matrix or
in \code{\link[ICAMS]{ICAMS}} format.}

\item{seedNumber}{A random seed that ensures ensures reproducible
results.}

\item{K.guess}{Suggested initial value of the number of
raw clusters. Usually, the number of raw clusters is
roughly twice the number
of extracted signatures.
Passed to hdpx::dp_activate as argument initcc.}

\item{multi.types}{A logical scalar or
a character vector.

If \code{FALSE}, The HDP analysis
will regard all input spectra as one tumor type, and the
HDP structure will have one parent node for all tumors.

If \code{TRUE}, Sample IDs in \code{input.catalog} must have the
form \emph{sample_type}::\emph{sample_id}.

If a character vector, then its length must be \code{ncol(input.catalog}),
and each value is the sample type of the corresponding column in \code{input.catalog},
e.g. \code{c(rep("Type-A", 23), rep("Type-B", 10))} for 23 Type-A samples and 10 Type-B
samples.

If not \code{FALSE}, HDP will have one parent node for each sample type
and one grandparent node.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}

\item{gamma.alpha}{Shape parameter of
the gamma distribution prior for the Dirichlet process concentration
parameters \eqn{\alpha_0} and all \eqn{\alpha_j} in
Figure B.1 of \itemize{
\item https://www.repository.cam.ac.uk/bitstream/handle/1810/275454/Roberts-2018-PhD.pdf
}}

\item{gamma.beta}{Inverse scale parameter (rate parameter) of
the gamma distribution prior for the Dirichlet process concentration
parameters:
\eqn{\beta_0} and all \eqn{\beta_j} in
Figure B.1 of \itemize{

\item https://www.repository.cam.ac.uk/bitstream/handle/1810/275454/Roberts-2018-PhD.pdf
}
We recommend gamma.alpha = 1 and gamma.beta = 20 for
single-base-substitution signature extraction;
gamma.alpha = 1 and gamma.beta = 50  for doublet-base-substitution
and indel signature extraction}
}
\value{
Invisibly, an \code{\link[hdpx]{hdpState-class}} object
as returned from \code{\link[hdpx]{dp_activate}}.
}
\description{
Generate an HDP Gibbs sampling chain from a spectra catalog.
}
\keyword{internal}
