% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetUpAndPosterior.R,
%   R/sketchCopyOfSetUpAndPosterior.R
\name{SetUpAndPosterior}
\alias{SetUpAndPosterior}
\title{Generate HDP gibs sampling chains from a spectra catalog.}
\usage{
SetUpAndPosterior(
  input.catalog,
  CPU.cores = 1,
  seedNumber = 1,
  K.guess,
  multi.types = FALSE,
  verbose = TRUE,
  num.posterior = 4,
  post.burnin = 4000,
  post.n = 50,
  post.space = 50,
  post.cpiter = 3,
  post.verbosity = 0,
  checkpoint.aft.post = NULL,
  stop.after.hdp.posterior = NULL
)

SetUpAndPosterior(
  input.catalog,
  CPU.cores = 1,
  seedNumber = 1,
  K.guess,
  multi.types = FALSE,
  verbose = TRUE,
  num.posterior = 4,
  post.burnin = 4000,
  post.n = 50,
  post.space = 50,
  post.cpiter = 3,
  post.verbosity = 0,
  checkpoint.aft.post = NULL,
  stop.after.hdp.posterior = NULL
)
}
\arguments{
\item{input.catalog}{Input spectra catalog as a matrix or
in \code{\link[ICAMS]{ICAMS}} format.}

\item{CPU.cores}{Number of CPUs to use in running
\code{\link[hdpx]{hdp_posterior}}; this is used to parallelize
running the posterior sampling chains, so there is no
point in making this larger than \code{num.posterior}.}

\item{seedNumber}{An integer that is used to generate separate
random seeds for each call to \code{\link[hdpx]{dp_activate}},
and each call of \code{\link[hdpx]{hdp_posterior}}; please see the code
on how this is done. But repeated calls with same value of
\code{seedNumber} and other inputs should produce the same results.}

\item{K.guess}{Suggested initial value of the number of
signatures, passed to \code{\link[hdpx]{dp_activate}} as
\code{initcc}.}

\item{multi.types}{A logical scalar or
a character vector.
If \code{FALSE}, hdp will regard all input spectra as one tumor type.

If \code{TRUE}, hdp will infer tumor types based on the string before "::" in their names.
e.g. tumor type for "SA.Syn.Ovary-AdenoCA::S.500" would be "SA.Syn.Ovary-AdenoCA"

If \code{multi.types} is a character vector, then it should be of the same length
as the number of columns in \code{input.catalog}, and each value is the
name of the tumor type of the corresponding column in \code{input.catalog},
e.g. \code{c("SA.Syn.Ovary-AdenoCA", "SA.Syn.Ovary-AdenoCA", "SA.Syn.Kidney-RCC")}.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}

\item{num.posterior}{Number of posterior sampling chains; can set to
1 for testing.}

\item{post.burnin}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{burnin}.}

\item{post.n}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{n}.}

\item{post.space}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{space}.}

\item{post.cpiter}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{cpiter}.}

\item{post.verbosity}{Pass to \code{\link[hdpx]{hdp_posterior}}
\code{verbosity}.}

\item{checkpoint.aft.post}{If non-\code{NULL}, a file path to checkpoint
the list of values returned from the calls to \code{\link[hdpx]{hdp_posterior}}
as a .Rdata file.}

\item{stop.after.hdp.posterior}{If non-\code{NULL}, then
a file path to checkpoint
the return value from the call to \code{\link[hdpx]{hdp_posterior}}
as a .Rdata file. The function will then also invisibly return
this value.}
}
\value{
Invisibly,
the clean
\code{chlist} (output of the hdp_posterior calls).

Invisibly,
the clean
\code{chlist} (output of the hdp_posterior calls).
}
\description{
Generate HDP gibs sampling chains from a spectra catalog.

Generate HDP gibs sampling chains from a spectra catalog.
}
