% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetupAndPosterior.R
\name{SetupAndPosterior}
\alias{SetupAndPosterior}
\title{Generate an HDP Gibbs sampling chain from a spectra catalog.}
\usage{
SetupAndPosterior(
  input.catalog,
  seedNumber = 1,
  K.guess,
  multi.types = FALSE,
  verbose = TRUE,
  burnin = 5000,
  post.n = 50,
  post.space = 50,
  post.cpiter = 3,
  post.verbosity = 0,
  gamma.alpha = 1,
  gamma.beta = 20,
  gamma0.alpha = gamma.alpha,
  gamma0.beta = gamma.beta,
  checkpoint.1.chain = TRUE,
  burnin.multiplier = 2,
  burnin.checkpoint = TRUE,
  prior.sigs = NULL,
  prior.pseudoc = NULL
)
}
\arguments{
\item{input.catalog}{Input spectra catalog as a matrix or
in \code{\link[ICAMS]{ICAMS}} format.}

\item{seedNumber}{A random seeds passed to \code{\link[hdpx]{dp_activate}}.}

\item{K.guess}{Suggested initial value of the number of
clusters. Usually, the number of clusters is two times of the number
of extracted signatures. Passed to \code{\link[hdpx]{dp_activate}} as
\code{initcc}.}

\item{multi.types}{A logical scalar or
a character vector.
If \code{FALSE}, The HDP analysis
will regard all input spectra as one tumor type.
HDP structure as one parent node for all tumors

If \code{TRUE}, the HDP analysis
will infer tumor types based on the string before "::" in their names.
e.g. tumor type for "Ovary-AdenoCA::S.500" would be "Ovary-AdenoCA"
HDP structure as a grandparent node for whole data and
one parent node for each tumor type

If \code{multi.types} is a character vector, then it should be of the same length
as the number of columns in \code{input.catalog}, and each value is the
name of the tumor type of the corresponding column in \code{input.catalog}.

If not \code{FALSE}, HDP will give a parent node for each tumor type and a grandparent node
for the whole dataset.

e.g. \code{c("Ovary-AdenoCA", "Kidney-RCC")}.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}

\item{burnin}{Pass to \code{\link[hdpx]{hdp_burnin}}
\code{burnin}. The number of burn-in iterations}

\item{post.n}{Pass to \code{\link[hdpx]{hdp_posterior_sample}}
\code{n}.The number of posterior samples to collect.}

\item{post.space}{Pass to \code{\link[hdpx]{hdp_posterior_sample}}
\code{space}. The number of iterations between collected samples.}

\item{post.cpiter}{Pass to \code{\link[hdpx]{hdp_posterior_sample}} and
\code{\link[hdpx]{hdp_burnin}} \code{cpiter}.The number of iterations of concentration
parameter sampling to perform after each iteration}

\item{post.verbosity}{Pass to \code{\link[hdpx]{hdp_posterior_sample}}
\code{verbosity}. Verbosity of debugging statements.
No need to change unless for development purpose}

\item{gamma.alpha}{Shape parameter of
the gamma distribution prior for the Dirichlet process concentration
parameters; in this
function the gamma distributions for all Dirichlet processes,
except possibly the top level process, are the same.}

\item{gamma.beta}{Inverse scale parameter (rate parameter) of
the gamma distribution prior for the Dirichlet process concentration
parameters; in this
function the gamma distributions for all Dirichlet processes, except
possibly the top level process, are the same.

We recommend gamma.alpha = 1 and gamma.beta = 20 for single-base-substitution signatures extraction;
gamma.alpha = 1 and gamma.beta = 50  for doublet-base-substitution/INDEL signature extraction}

\item{gamma0.alpha}{See figure B.1 from Nicola Robert's thesis.
The shape parameter (\eqn{\alpha_0}) of the gamma
distribution priors for
the Dirichlet process concentration parameters  (\eqn{\gamma_0})
for \eqn{G_0}.}

\item{gamma0.beta}{See figure B.1 from Nicola Robert's thesis.
Inverse scale parameter (rate parameter, \eqn{\beta_0}) of the gamma
distribution priors for
the Dirichlet process concentration parameters (\eqn{\gamma_0})
for \eqn{G_0}.}

\item{checkpoint.1.chain}{If \code{TRUE} checkpoint the sample
chain to current working directory, in a file called
sample.chain.\emph{seed_number}.Rdata.}

\item{burnin.multiplier}{A checkpoint setting. \code{burnin.multiplier} rounds of \code{burnin} iterations will be run.
After each round, a burn-in chain will be save for checkpoint.
A total number of 10,000 iterations is recommended for most analysis. However, number of iterations
can be adjusted based on the size of dataset. According to our experience, 20,000 iterations are
needed when analyzing all PCAWG7 genomes (2,780 samples).
The burnin can be continued from a checkpoint file with \code{\link{ExtendBurnin}}.}

\item{burnin.checkpoint}{Default is False. If True, a checkpoint for burnin will be created.}

\item{prior.sigs}{A matrix containing prior signatures.}

\item{prior.pseudoc}{A numeric list. Pesudo counts of each prior signature. Recommended is 1000. In practice,
it may be advisable to put lower weights on prior signatures that you do not expect
to be present in your dataset, or even exclude some priors entirely.}
}
\value{
Invisibly, an \code{\link[hdpx]{hdpSampleChain-class}} object
as returned from \code{\link[hdpx]{hdp_posterior}}.
}
\description{
Generate an HDP Gibbs sampling chain from a spectra catalog.
}
